@@ -9,8 +9,11 @@ import (
 
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/exe"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/logger"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/packagerepo/repocloner/rpmrepocloner"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/pkggraph"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/pkgjson"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/timestamp"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/pkg/profile"
 
 	"gopkg.in/alecthomas/kingpin.v2"
 )
@@ -20,10 +23,26 @@ var (
 	input  = exe.InputFlag(app, "Input json listing all local SRPMs")
 	output = exe.OutputFlag(app, "Output file to export the graph to")
 
-	logFile          = exe.LogFileFlag(app)
-	logLevel         = exe.LogLevelFlag(app)
-	strictGoals      = app.Flag("strict-goals", "Don't allow missing goal packages").Bool()
-	strictUnresolved = app.Flag("strict-unresolved", "Don't allow missing unresolved packages").Bool()
+	logFile               = exe.LogFileFlag(app)
+	logLevel              = exe.LogLevelFlag(app)
+	profFlags             = exe.SetupProfileFlags(app)
+	strictGoals           = app.Flag("strict-goals", "Don't allow missing goal packages").Bool()
+	strictUnresolved      = app.Flag("strict-unresolved", "Don't allow missing unresolved packages").Bool()
+	timestampFile         = app.Flag("timestamp-file", "File that stores timestamps for this program.").String()
+	usePMCtoResolveCycles = app.Flag("usePMCtoresolvecycles", "Cycles will be resolved by downloading rpm packages from PMC if locally unavailable").Bool()
+	tlsClientCert         = app.Flag("tls-cert", "TLS client certificate to use when downloading files.").String()
+	tlsClientKey          = app.Flag("tls-key", "TLS client key to use when downloading files.").String()
+
+	resolveCyclesFromUpstream     = app.Flag("resolve-cycles-from-upstream", "Let grapher resolve cycles by marking rpms available in repo as remote").Bool()
+	outDir                        = exe.OutputDirFlag(app, "Directory to download packages into.")
+	existingRpmsDir               = app.Flag("rpm-dir", "Directory that contains already built RPMs. Should contain top level directories for architecture.").Required().ExistingDir()
+	existingToolchainRpmDir       = app.Flag("toolchain-rpms-dir", "Directory that contains already built toolchain RPMs. Should contain top level directories for architecture.").Required().ExistingDir()
+	tmpDir                        = app.Flag("tmp-dir", "Directory to store temporary files while downloading.").String()
+	workerTar                     = app.Flag("tdnf-worker", "Full path to worker_chroot.tar.gz").Required().ExistingFile()
+	repoFiles                     = app.Flag("repo-file", "Full path to a repo file").Required().ExistingFiles()
+	usePreviewRepo                = app.Flag("use-preview-repo", "Pull packages from the upstream preview repo").Bool()
+	disableDefaultRepos           = app.Flag("disable-default-repos", "Disable pulling packages from PMC repos").Bool()
+	ignoreVersionToResolveSelfDep = app.Flag("ignore-version-to-resolve-selfdep", "Ignore package version while downloading package from upstream when resolving cycle").Bool()
 
 	depGraph = pkggraph.NewPkgGraph()
 )
@@ -33,9 +52,16 @@ func main() {
 
 	app.Version(exe.ToolkitVersion)
 	kingpin.MustParse(app.Parse(os.Args[1:]))
-
-	var err error
 	logger.InitBestEffort(*logFile, *logLevel)
+
+	prof, err := profile.StartProfiling(profFlags)
+	if err != nil {
+		logger.Log.Warnf("Could not start profiling: %s", err)
+	}
+	defer prof.StopProfiler()
+
+	timestamp.BeginTiming("grapher", *timestampFile)
+	defer timestamp.CompleteTiming()
 
 	localPackages := pkgjson.PackageRepo{}
 	err = localPackages.ParsePackageJSON(*input)
@@ -43,19 +69,37 @@ func main() {
 		logger.Log.Panic(err)
 	}
 
+	depGraph := pkggraph.NewPkgGraph()
 	err = populateGraph(depGraph, &localPackages)
 	if err != nil {
 		logger.Log.Panic(err)
 	}
 
 	// Add a default "ALL" goal to build everything local
-	_, err = depGraph.AddGoalNode(goalNodeName, nil, *strictGoals)
+	_, err = depGraph.AddGoalNode(goalNodeName, nil, nil, *strictGoals)
 	if err != nil {
 		logger.Log.Panic(err)
 	}
 
+	var cloner *rpmrepocloner.RpmRepoCloner = nil
+	if *resolveCyclesFromUpstream {
+		cloner, err = rpmrepocloner.ConstructCloner(*outDir, *tmpDir, *workerTar, *existingRpmsDir, *existingToolchainRpmDir, *tlsClientCert, *tlsClientKey, *repoFiles)
+		if err != nil {
+			logger.Log.Panic(err)
+		}
+		enabledRepos := rpmrepocloner.RepoFlagAll
+		if !*usePreviewRepo {
+			enabledRepos = enabledRepos & ^rpmrepocloner.RepoFlagPreview
+		}
+		if *disableDefaultRepos {
+			enabledRepos = enabledRepos & ^rpmrepocloner.RepoFlagMarinerDefaults
+		}
+		cloner.SetEnabledRepos(enabledRepos)
+
+		defer cloner.Close()
+	}
 	logger.Log.Info("Running cycle resolution to fix any cycles in the dependency graph")
-	err = depGraph.MakeDAG()
+	err = depGraph.MakeDAGUsingUpstreamRepos(*resolveCyclesFromUpstream, *ignoreVersionToResolveSelfDep, cloner)
 	if err != nil {
 		logger.Log.Panic(err)
 	}
@@ -88,57 +132,74 @@ func addUnresolvedPackage(g *pkggraph.PkgGraph, pkgVer *pkgjson.PackageVer) (new
 	}
 
 	// Create a new node
-	newRunNode, err = g.AddPkgNode(pkgVer, pkggraph.StateUnresolved, pkggraph.TypeRemote, "<NO_SRPM_PATH>", "<NO_RPM_PATH>", "<NO_SPEC_PATH>", "<NO_SOURCE_PATH>", "<NO_ARCHITECTURE>", "<NO_REPO>")
+	newRunNode, err = g.AddRemoteUnresolvedNode(pkgVer)
 	if err != nil {
 		return
 	}
 
-	logger.Log.Infof("Adding unresolved node %s\n", newRunNode.FriendlyName())
+	logger.Log.Infof("Adding unresolved node '%s'.", newRunNode.FriendlyName())
 
 	return
 }
 
-// addNodesForPackage creates a "Run" and "Build" node for the package described
-// in the PackageVer structure. Returns pointers to the build and run Nodes
+// addNodesForPackage creates a "Run", "Build", and "Test" node for the package described
+// in the Package structure. Returns pointers to the build and run Nodes
 // created, or an error if one of the nodes could not be created.
-func addNodesForPackage(g *pkggraph.PkgGraph, pkgVer *pkgjson.PackageVer, pkg *pkgjson.Package) (newRunNode *pkggraph.PkgNode, newBuildNode *pkggraph.PkgNode, err error) {
-	nodes, err := g.FindExactPkgNodeFromPkg(pkgVer)
+func addNodesForPackage(g *pkggraph.PkgGraph, pkg *pkgjson.Package) (foundDuplicate bool, err error) {
+	var (
+		newRunNode   *pkggraph.PkgNode
+		newBuildNode *pkggraph.PkgNode
+		newTestNode  *pkggraph.PkgNode
+	)
+
+	nodes, err := g.FindExactPkgNodeFromPkg(pkg.Provides)
 	if err != nil {
 		return
 	}
+
 	if nodes != nil {
-		logger.Log.Warnf(`Duplicate package name for package %+v read from SRPM "%s" (Previous: %+v)`, pkgVer, pkg.SrpmPath, nodes.RunNode)
-		err = nil
-		if nodes.RunNode != nil {
-			newRunNode = nodes.RunNode
-		}
-		if nodes.BuildNode != nil {
-			newBuildNode = nodes.BuildNode
-		}
+		logger.Log.Warnf(`Skipping duplicate package name for package %+v read from SRPM "%s". Original: %+v.`, pkg.Provides, pkg.SrpmPath, nodes.RunNode)
+		foundDuplicate = true
+		return
 	}
 
-	if newRunNode == nil {
-		// Add "Run" node
-		newRunNode, err = g.AddPkgNode(pkgVer, pkggraph.StateMeta, pkggraph.TypeRun, pkg.SrpmPath, pkg.RpmPath, pkg.SpecPath, pkg.SourceDir, pkg.Architecture, "<LOCAL>")
-		logger.Log.Debugf("Adding run node %s with id %d\n", newRunNode.FriendlyName(), newRunNode.ID())
-		if err != nil {
-			return
-		}
+	newRunNode, err = g.AddPkgNode(pkg.Provides, pkggraph.StateMeta, pkggraph.TypeLocalRun, pkg.SrpmPath, pkg.RpmPath, pkg.SpecPath, pkg.SourceDir, pkg.Architecture, pkggraph.LocalRepo)
+	if err != nil {
+		return
 	}
+	logger.Log.Debugf("Adding run node '%s' with id %d.", newRunNode.FriendlyName(), newRunNode.ID())
 
-	if newBuildNode == nil {
-		// Add "Build" node
-		newBuildNode, err = g.AddPkgNode(pkgVer, pkggraph.StateBuild, pkggraph.TypeBuild, pkg.SrpmPath, pkg.RpmPath, pkg.SpecPath, pkg.SourceDir, pkg.Architecture, "<LOCAL>")
-		logger.Log.Debugf("Adding build node %s with id %d\n", newBuildNode.FriendlyName(), newBuildNode.ID())
-		if err != nil {
-			return
-		}
+	newBuildNode, err = g.AddPkgNode(pkg.Provides, pkggraph.StateBuild, pkggraph.TypeLocalBuild, pkg.SrpmPath, pkg.RpmPath, pkg.SpecPath, pkg.SourceDir, pkg.Architecture, pkggraph.LocalRepo)
+	if err != nil {
+		return
 	}
+	logger.Log.Debugf("Adding build node '%s' with id %d.", newBuildNode.FriendlyName(), newBuildNode.ID())
 
-	// A "run" node has an implicit dependency on its coresponding "build" node, encode that here.
+	// A "run" node has an implicit dependency on its corresponding "build" node, encode that here.
 	err = g.AddEdge(newRunNode, newBuildNode)
 	if err != nil {
-		logger.Log.Errorf("Adding edge failed for %+v", pkgVer)
+		logger.Log.Errorf("Adding run -> build edge failed for %+v", pkg.Provides)
+		return
+	}
+
+	if !pkg.RunTests {
+		logger.Log.Debugf("Skipping adding a test node for package %+v", pkg)
+		return
+	}
+
+	newTestNode, err = g.AddPkgNode(pkg.Provides, pkggraph.StateBuild, pkggraph.TypeTest, pkg.SrpmPath, pkggraph.NoRPMPath, pkg.SpecPath, pkg.SourceDir, pkg.Architecture, pkggraph.LocalRepo)
+	if err != nil {
+		return
+	}
+	logger.Log.Debugf("Adding test node '%s' with id %d.", newTestNode.FriendlyName(), newTestNode.ID())
+
+	// A "test" node has a dependency on its corresponding "build" node. This dependency is required
+	// to guarantee we will first check if the build node needs to be built or not before we make
+	// any decisions about running the tests.
+	err = g.AddEdge(newTestNode, newBuildNode)
+	if err != nil {
+		logger.Log.Errorf("Adding test -> build edge failed for %+v", pkg.Provides)
+		return
 	}
 
 	return
@@ -176,8 +237,8 @@ func addSingleDependency(g *pkggraph.PkgGraph, packageNode *pkggraph.PkgNode, de
 	// Creating these edges may cause non-problematic cycles that can significantly increase memory usage and runtime during cycle resolution.
 	// If there are enough of these cycles it can exhaust the system's memory when resolving them.
 	// - Only check run nodes. If a build node has a reflexive cycle then it cannot be built without a bootstrap version.
-	if packageNode.Type == pkggraph.TypeRun &&
-		dependentNode.Type == pkggraph.TypeRun &&
+	if packageNode.Type == pkggraph.TypeLocalRun &&
+		dependentNode.Type == pkggraph.TypeLocalRun &&
 		packageNode.RpmPath == dependentNode.RpmPath {
 
 		logger.Log.Debugf("%+v requires %+v which is provided by the same RPM.", packageNode, dependentNode)
@@ -192,37 +253,24 @@ func addSingleDependency(g *pkggraph.PkgGraph, packageNode *pkggraph.PkgNode, de
 	return err
 }
 
-// addLocalPackage adds the package provided by the Package structure, and
-// updates the SRPM path name
-func addLocalPackage(g *pkggraph.PkgGraph, pkg *pkgjson.Package) error {
-	_, _, err := addNodesForPackage(g, pkg.Provides, pkg)
-	return err
-}
-
-// addDependencies adds edges for both build and runtime requirements for the
+// addDependencies adds edges for run-, build-, and test-time requirements for the
 // package described in the Package structure. Returns an error if the edges
 // could not be created.
 func addPkgDependencies(g *pkggraph.PkgGraph, pkg *pkgjson.Package) (dependenciesAdded int, err error) {
-	provide := pkg.Provides
-	runDependencies := pkg.Requires
-	buildDependencies := pkg.BuildRequires
-
 	// Find the current node in the lookup list.
 	logger.Log.Debugf("Adding dependencies for package %s", pkg.SrpmPath)
-	nodes, err := g.FindExactPkgNodeFromPkg(provide)
+	nodes, err := g.FindExactPkgNodeFromPkg(pkg.Provides)
 	if err != nil {
 		return
 	}
 	if nodes == nil {
 		return dependenciesAdded, fmt.Errorf("can't add dependencies to a missing package %+v", pkg)
 	}
-	runNode := nodes.RunNode
-	buildNode := nodes.BuildNode
 
-	// For each run time and build time dependency, add the edges
+	// For each run-, build-, and test-time dependency, add the edges
 	logger.Log.Tracef("Adding run dependencies")
-	for _, dependency := range runDependencies {
-		err = addSingleDependency(g, runNode, dependency)
+	for _, dependency := range pkg.Requires {
+		err = addSingleDependency(g, nodes.RunNode, dependency)
 		if err != nil {
 			logger.Log.Errorf("Unable to add run-time dependencies for %+v", pkg)
 			return
@@ -231,10 +279,25 @@ func addPkgDependencies(g *pkggraph.PkgGraph, pkg *pkgjson.Package) (dependencie
 	}
 
 	logger.Log.Tracef("Adding build dependencies")
-	for _, dependency := range buildDependencies {
-		err = addSingleDependency(g, buildNode, dependency)
+	for _, dependency := range pkg.BuildRequires {
+		err = addSingleDependency(g, nodes.BuildNode, dependency)
 		if err != nil {
 			logger.Log.Errorf("Unable to add build-time dependencies for %+v", pkg)
+			return
+		}
+		dependenciesAdded++
+	}
+
+	if nodes.TestNode == nil {
+		logger.Log.Debugf("No test node for package %+v, skipping test dependencies", pkg)
+		return
+	}
+
+	logger.Log.Tracef("Adding test dependencies")
+	for _, dependency := range pkg.TestRequires {
+		err = addSingleDependency(g, nodes.TestNode, dependency)
+		if err != nil {
+			logger.Log.Errorf("Unable to add test-time dependencies for %+v", pkg)
 			return
 		}
 		dependenciesAdded++
@@ -246,34 +309,46 @@ func addPkgDependencies(g *pkggraph.PkgGraph, pkg *pkgjson.Package) (dependencie
 // populateGraph adds all the data contained in the PackageRepo structure into
 // the graph.
 func populateGraph(graph *pkggraph.PkgGraph, repo *pkgjson.PackageRepo) (err error) {
+	timestamp.StartEvent("populating graph", nil)
+	defer timestamp.StopEvent(nil)
+
 	packages := repo.Repo
+
+	timestamp.StartEvent("add package node", nil)
 
 	// Scan and add each package we know about
 	logger.Log.Infof("Adding all packages from %s", *input)
-	// NOTE: range iterates by value, not reference. Manually access slice
-	for idx := range packages {
-		pkg := packages[idx]
-		err = addLocalPackage(graph, pkg)
+	uniquePackages := make(map[*pkgjson.Package]bool)
+	for _, pkg := range packages {
+		foundDuplicate, err := addNodesForPackage(graph, pkg)
 		if err != nil {
 			logger.Log.Errorf("Failed to add local package %+v", pkg)
 			return err
 		}
+
+		if !foundDuplicate {
+			uniquePackages[pkg] = true
+		}
 	}
 	logger.Log.Infof("\tAdded %d packages", len(packages))
+
+	timestamp.StopEvent(nil) // add package nodes
+	timestamp.StartEvent("add dependencies", nil)
 
 	// Rescan and add all the dependencies
 	logger.Log.Infof("Adding all dependencies from %s", *input)
 	dependenciesAdded := 0
-	for idx := range packages {
-		pkg := packages[idx]
-		num, err := addPkgDependencies(graph, pkg)
+	for uniquePkg := range uniquePackages {
+		num, err := addPkgDependencies(graph, uniquePkg)
 		if err != nil {
-			logger.Log.Errorf("Failed to add dependency %+v", pkg)
+			logger.Log.Errorf("Failed to add dependency %+v", uniquePkg)
 			return err
 		}
 		dependenciesAdded += num
 	}
 	logger.Log.Infof("\tAdded %d dependencies", dependenciesAdded)
+
+	timestamp.StopEvent(nil) // add dependencies
 
 	return err
 }
