@@ -13,13 +13,14 @@ import (
 
 // KernelCommandLine holds extra command line parameters which can be
 // added to the grub config file.
-// - ImaPolicy: A list of IMA policies which will be used together
-// - ExtraCommandLine: Arbitrary parameters which will be appended to the
-//   end of the kernel command line
+//   - ImaPolicy: A list of IMA policies which will be used together
+//   - ExtraCommandLine: Arbitrary parameters which will be appended to the
+//     end of the kernel command line
 type KernelCommandLine struct {
 	CGroup           CGroup      `json:"CGroup"`
 	ImaPolicy        []ImaPolicy `json:"ImaPolicy"`
 	SELinux          SELinux     `json:"SELinux"`
+	EnableFIPS       bool        `json:"EnableFIPS"`
 	ExtraCommandLine string      `json:"ExtraCommandLine"`
 }
 
@@ -49,7 +50,7 @@ func (k *KernelCommandLine) IsValid() (err error) {
 
 	// A character needs to be set aside for use as the sed delimiter, make sure it isn't included in the provided string
 	if strings.Contains(k.ExtraCommandLine, k.GetSedDelimeter()) {
-		return fmt.Errorf("ExtraCommandLine contains character %s which is reserved for use by sed", k.GetSedDelimeter())
+		return fmt.Errorf("the 'ExtraCommandLine' field contains the %s character which is reserved for use by sed", k.GetSedDelimeter())
 	}
 
 	return
