@@ -5,6 +5,7 @@ package configuration
 
 import (
 	"os"
+	"path/filepath"
 	"testing"
 
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/file"
@@ -134,28 +135,12 @@ func TestShouldFailParsingInvalidDevice_Network(t *testing.T) {
 }
 
 func TestShouldPassCreatingNetworkFile_Network(t *testing.T) {
-	const (
-		networkFileDir  = "/etc/systemd/network"
-		testNetworkFile = "/etc/systemd/network/10-static-eth1.network"
-	)
-
-	// Some systems may not have systemd-networkd service configured, and thus
-	// /etc/systemd/network may not exist. For this test case, manually create this directory
-	// if it does not exist
-	exists, err := file.DirExists(networkFileDir)
-	assert.NoError(t, err)
-	if !exists {
-		err = os.Mkdir(networkFileDir, os.ModePerm)
-		assert.NoError(t, err)
-		t.Cleanup(func() {
-			err = os.RemoveAll(networkFileDir)
-			assert.NoError(t, err)
-		})
-	}
+	testNetworkFileDir := t.TempDir()
+	testNetworkFile := filepath.Join(testNetworkFileDir, "10-static-eth1.network")
 
 	testNetwork := validNetworks[0]
 
-	err = createNetworkConfigFile(nil, testNetwork, "eth1")
+	err := createNetworkConfigFile(nil, testNetwork, "eth1", testNetworkFileDir)
 	assert.NoError(t, err)
 	t.Cleanup(func() {
 		os.Remove(testNetworkFile)
