@@ -16,53 +16,41 @@ import (
 
 // SystemConfig defines how each system present on the image is supposed to be configured.
 type SystemConfig struct {
-	IsDefault            bool               `json:"IsDefault"`
-	IsKickStartBoot      bool               `json:"IsKickStartBoot"`
-	IsIsoInstall         bool               `json:"IsIsoInstall"`
-	BootType             string             `json:"BootType"`
-	Hostname             string             `json:"Hostname"`
-	Name                 string             `json:"Name"`
-	PackageLists         []string           `json:"PackageLists"`
-	Packages             []string           `json:"Packages"`
-	KernelOptions        map[string]string  `json:"KernelOptions"`
-	KernelCommandLine    KernelCommandLine  `json:"KernelCommandLine"`
-	AdditionalFiles      map[string]string  `json:"AdditionalFiles"`
-	PartitionSettings    []PartitionSetting `json:"PartitionSettings"`
-	PreInstallScripts    []InstallScript    `json:"PreInstallScripts"`
-	PostInstallScripts   []InstallScript    `json:"PostInstallScripts"`
-	FinalizeImageScripts []InstallScript    `json:"FinalizeImageScripts"`
-	Networks             []Network          `json:"Networks"`
-	PackageRepos         []PackageRepo      `json:"PackageRepos"`
-	Groups               []Group            `json:"Groups"`
-	Users                []User             `json:"Users"`
-	Encryption           RootEncryption     `json:"Encryption"`
-	RemoveRpmDb          bool               `json:"RemoveRpmDb"`
-	ReadOnlyVerityRoot   ReadOnlyVerityRoot `json:"ReadOnlyVerityRoot"`
-	HidepidDisabled      bool               `json:"HidepidDisabled"`
+	IsDefault            bool                      `json:"IsDefault"`
+	IsKickStartBoot      bool                      `json:"IsKickStartBoot"`
+	IsIsoInstall         bool                      `json:"IsIsoInstall"`
+	BootType             string                    `json:"BootType"`
+	Hostname             string                    `json:"Hostname"`
+	Name                 string                    `json:"Name"`
+	PackageLists         []string                  `json:"PackageLists"`
+	Packages             []string                  `json:"Packages"`
+	KernelOptions        map[string]string         `json:"KernelOptions"`
+	KernelCommandLine    KernelCommandLine         `json:"KernelCommandLine"`
+	AdditionalFiles      map[string]FileConfigList `json:"AdditionalFiles"`
+	PartitionSettings    []PartitionSetting        `json:"PartitionSettings"`
+	PreInstallScripts    []InstallScript           `json:"PreInstallScripts"`
+	PostInstallScripts   []InstallScript           `json:"PostInstallScripts"`
+	FinalizeImageScripts []InstallScript           `json:"FinalizeImageScripts"`
+	Networks             []Network                 `json:"Networks"`
+	PackageRepos         []PackageRepo             `json:"PackageRepos"`
+	Groups               []Group                   `json:"Groups"`
+	Users                []User                    `json:"Users"`
+	Encryption           RootEncryption            `json:"Encryption"`
+	RemoveRpmDb          bool                      `json:"RemoveRpmDb"`
+	ReadOnlyVerityRoot   ReadOnlyVerityRoot        `json:"ReadOnlyVerityRoot"`
+	HidepidDisabled      bool                      `json:"HidepidDisabled"`
 }
 
 // GetRootPartitionSetting returns a pointer to the partition setting describing the disk which
 // will be mounted at "/", or nil if no partition is found
 func (s *SystemConfig) GetRootPartitionSetting() (rootPartitionSetting *PartitionSetting) {
-	for i, p := range s.PartitionSettings {
-		if p.MountPoint == "/" {
-			// We want to reference the actual object in the slice
-			return &s.PartitionSettings[i]
-		}
-	}
-	return nil
+	return FindRootPartitionSetting(s.PartitionSettings)
 }
 
 // GetMountpointPartitionSetting will search the system configuration for the partition setting
 // corresponding to a mount point.
 func (s *SystemConfig) GetMountpointPartitionSetting(mountPoint string) (partitionSetting *PartitionSetting) {
-	for i, p := range s.PartitionSettings {
-		if p.MountPoint == mountPoint {
-			// We want to reference the actual object in the slice
-			return &s.PartitionSettings[i]
-		}
-	}
-	return nil
+	return FindMountpointPartitionSetting(s.PartitionSettings, mountPoint)
 }
 
 // IsValid returns an error if the SystemConfig is not valid
@@ -80,7 +68,7 @@ func (s *SystemConfig) IsValid() (err error) {
 
 	// Validate BootType
 	if s.BootType != "efi" && s.BootType != "legacy" && s.BootType != "none" && s.BootType != "" {
-		return fmt.Errorf("invalid [BootType]: %s. Expecting values of either 'efi', 'legacy', 'none' or empty string.", s.BootType)
+		return fmt.Errorf("invalid [BootType]: %s. Expecting values of either 'efi', 'legacy', 'none' or empty string", s.BootType)
 	}
 
 	if len(s.PackageLists) == 0 && len(s.Packages) == 0 {
@@ -147,6 +135,13 @@ func (s *SystemConfig) IsValid() (err error) {
 
 	if err = s.KernelCommandLine.IsValid(); err != nil {
 		return fmt.Errorf("invalid [KernelCommandLine]: %w", err)
+	}
+
+	for srcFile, fileConfigList := range s.AdditionalFiles {
+		err = fileConfigList.IsValid()
+		if err != nil {
+			return fmt.Errorf("invalid [AdditionalFiles]: (%s): %w", srcFile, err)
+		}
 	}
 
 	// Validate that PackageRepos do not contain duplicate package repo name
