@@ -163,11 +163,11 @@ func TestShouldFailParsingBadKernelCommandLine_SystemConfig(t *testing.T) {
 
 	err := badKernelCommandConfig.IsValid()
 	assert.Error(t, err)
-	assert.Equal(t, "invalid [KernelCommandLine]: ExtraCommandLine contains character ` which is reserved for use by sed", err.Error())
+	assert.Equal(t, "invalid [KernelCommandLine]: the 'ExtraCommandLine' field contains the ` character which is reserved for use by sed", err.Error())
 
 	err = remarshalJSON(badKernelCommandConfig, &checkedSystemConfig)
 	assert.Error(t, err)
-	assert.Equal(t, "failed to parse [SystemConfig]: failed to parse [KernelCommandLine]: ExtraCommandLine contains character ` which is reserved for use by sed", err.Error())
+	assert.Equal(t, "failed to parse [SystemConfig]: failed to parse [KernelCommandLine]: the 'ExtraCommandLine' field contains the ` character which is reserved for use by sed", err.Error())
 }
 
 func TestShouldFailParsingInvalidHostName_SystemConfig(t *testing.T) {
@@ -372,5 +372,14 @@ func TestShouldFailParsingUnexpectedBootType_SystemConfig(t *testing.T) {
 
 	err := invalidBootTypeConfig.IsValid()
 	assert.Error(t, err)
-	assert.Equal(t, "invalid [BootType]: uefi. Expecting values of either 'efi', 'legacy', 'none' or empty string.", err.Error())
+	assert.Equal(t, "invalid [BootType]: uefi. Expecting values of either 'efi', 'legacy', 'none' or empty string", err.Error())
+}
+
+func TestShouldFaiInvalidAdditionalFiles_SystemConfig(t *testing.T) {
+	systemConfig := validSystemConfig
+	systemConfig.AdditionalFiles = map[string]FileConfigList{"a.txt": {}}
+
+	err := systemConfig.IsValid()
+	assert.Error(t, err)
+	assert.Equal(t, "invalid [AdditionalFiles]: (a.txt): list is empty", err.Error())
 }
