@@ -17,6 +17,8 @@ import (
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/file"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/logger"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/safechroot"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/timestamp"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/pkg/profile"
 
 	"gopkg.in/alecthomas/kingpin.v2"
 )
@@ -31,10 +33,13 @@ var (
 	assets          = app.Flag("assets", "Path to assets directory.").ExistingDir()
 	baseDirPath     = app.Flag("base-dir", "Base directory for relative file paths from the config. Defaults to config's directory.").ExistingDir()
 	outputDir       = app.Flag("output-dir", "Path to directory to place final image.").ExistingDir()
+	imgContentFile  = app.Flag("output-image-contents", "File that stores list of packages used to compose the image.").String()
 	liveInstallFlag = app.Flag("live-install", "Enable to perform a live install to the disk specified in config file.").Bool()
 	emitProgress    = app.Flag("emit-progress", "Write progress updates to stdout, such as percent complete and current action.").Bool()
+	timestampFile   = app.Flag("timestamp-file", "File that stores timestamps for this program.").String()
 	logFile         = exe.LogFileFlag(app)
 	logLevel        = exe.LogLevelFlag(app)
+	profFlags       = exe.SetupProfileFlags(app)
 )
 
 const (
@@ -57,6 +62,8 @@ const (
 	// kickstartPartitionFile is the file that includes the partitioning schema used by
 	// kickstart installation
 	kickstartPartitionFile = "/tmp/part-include"
+
+	assetsMountPoint = "/installer"
 )
 
 func main() {
@@ -64,8 +71,16 @@ func main() {
 
 	app.Version(exe.ToolkitVersion)
 	kingpin.MustParse(app.Parse(os.Args[1:]))
-
 	logger.InitBestEffort(*logFile, *logLevel)
+
+	prof, err := profile.StartProfiling(profFlags)
+	if err != nil {
+		logger.Log.Warnf("Could not start profiling: %s", err)
+	}
+	defer prof.StopProfiler()
+
+	timestamp.BeginTiming("imager", *timestampFile)
+	defer timestamp.CompleteTiming()
 
 	if *emitProgress {
 		installutils.EnableEmittingProgress()
@@ -74,12 +89,12 @@ func main() {
 	// Parse Config
 	config, err := configuration.LoadWithAbsolutePaths(*configFile, *baseDirPath)
 	logger.PanicOnError(err, "Failed to load configuration file (%s) with base directory (%s)", *configFile, *baseDirPath)
-
 	// Currently only process 1 system config
 	systemConfig := config.SystemConfigs[defaultSystemConfig]
 
 	// Execute preinstall scripts and parse partitioning when performing kickstart installation
 	if systemConfig.IsKickStartBoot {
+		timestamp.StartEvent("applying kickstart", nil)
 		err = installutils.RunPreInstallScripts(systemConfig)
 		logger.PanicOnError(err, "Failed to preinstall scripts")
 
@@ -93,27 +108,29 @@ func main() {
 		if err != nil {
 			logger.PanicOnError(err, "Invalid image configuration: %s", err)
 		}
+		timestamp.StopEvent(nil) // applying kickstart
 	}
 
-	err = buildSystemConfig(systemConfig, config.Disks, *outputDir, *buildDir)
+	err = buildSystemConfig(systemConfig, config.Disks, *outputDir, *buildDir, *imgContentFile)
 	logger.PanicOnError(err, "Failed to build system configuration")
-
 }
 
-func buildSystemConfig(systemConfig configuration.SystemConfig, disks []configuration.Disk, outputDir, buildDir string) (err error) {
+func buildSystemConfig(systemConfig configuration.SystemConfig, disks []configuration.Disk, outputDir, buildDir string, imgContentFile string) (err error) {
 	logger.Log.Infof("Building system configuration (%s)", systemConfig.Name)
+	timestamp.StartEvent("building system config", nil)
+	defer timestamp.StopEvent(nil)
 
 	const (
-		assetsMountPoint    = "/installer"
-		localRepoMountPoint = "/mnt/cdrom/RPMS"
-		repoFileMountPoint  = "/etc/yum.repos.d"
-		setupRoot           = "/setuproot"
-		installRoot         = "/installroot"
-		rootID              = "rootfs"
-		defaultDiskIndex    = 0
-		defaultTempDiskName = "disk.raw"
-		existingChrootDir   = false
-		leaveChrootOnDisk   = false
+		localRepoMountPoint   = "/mnt/cdrom/RPMS"
+		repoFileMountPoint    = "/etc/yum.repos.d"
+		setupRoot             = "/setuproot"
+		installRoot           = "/installroot"
+		rootID                = "rootfs"
+		defaultDiskIndex      = 0
+		defaultTempDiskName   = "disk.raw"
+		existingChrootDir     = false
+		leaveChrootOnDisk     = false
+		marinerReleasePackage = "mariner-release"
 	)
 
 	var (
@@ -138,9 +155,16 @@ func buildSystemConfig(systemConfig configuration.SystemConfig, disks []configur
 		return
 	}
 
+	// Mariner images don't work appropriately when mariner-release is not installed.
+	// As a stopgap to this, mariner-release will now be added to all images regardless
+	// of presence in the CONFIG_FILE
+	packagesToInstall = append([]string{marinerReleasePackage}, packagesToInstall...)
+
 	isRootFS = len(systemConfig.PartitionSettings) == 0
 	if isRootFS {
 		logger.Log.Infof("Creating rootfs")
+		timestamp.StartEvent("creating rootfs", nil)
+		defer timestamp.StopEvent(nil)
 		additionalExtraMountPoints, additionalExtraDirectories, err := setupRootFS(outputDir, installRoot)
 		if err != nil {
 			return err
@@ -160,7 +184,8 @@ func buildSystemConfig(systemConfig configuration.SystemConfig, disks []configur
 			packagesToInstall = append([]string{kernelPkg}, packagesToInstall...)
 		}
 	} else {
-		logger.Log.Info("Creating raw disk in build directory")
+		timestamp.StartEvent("creating raw disk", nil)
+
 		diskConfig := disks[defaultDiskIndex]
 		diskDevPath, partIDToDevPathMap, partIDToFsTypeMap, isLoopDevice, encryptedRoot, readOnlyRoot, err = setupDisk(buildDir, defaultTempDiskName, *liveInstallFlag, diskConfig, systemConfig.Encryption, systemConfig.ReadOnlyVerityRoot)
 		if err != nil {
@@ -192,13 +217,15 @@ func buildSystemConfig(systemConfig configuration.SystemConfig, disks []configur
 
 		logger.Log.Infof("Selected (%s) for the kernel", kernelPkg)
 		packagesToInstall = append([]string{kernelPkg}, packagesToInstall...)
+		timestamp.StopEvent(nil) // creating raw disk
 	}
 
 	setupChrootDir := filepath.Join(buildDir, setupRoot)
 
 	// Create Parition to Mountpoint map
-	mountPointMap, mountPointToFsTypeMap, mountPointToMountArgsMap, diffDiskBuild := installutils.CreateMountPointPartitionMap(partIDToDevPathMap, partIDToFsTypeMap, systemConfig)
+	mountPointMap, mountPointToFsTypeMap, mountPointToMountArgsMap, diffDiskBuild := installutils.CreateMountPointPartitionMap(partIDToDevPathMap, partIDToFsTypeMap, systemConfig.PartitionSettings)
 	if diffDiskBuild {
+		timestamp.StartEvent("creating delta disk", nil)
 		mountPointToOverlayMap, err = installutils.UpdatePartitionMapWithOverlays(partIDToDevPathMap, partIDToFsTypeMap, mountPointMap, mountPointToFsTypeMap, mountPointToMountArgsMap, systemConfig)
 		// Schedule unmount of overlays after the upper layers are unmounted.
 		defer installutils.OverlayUnmount(mountPointToOverlayMap)
@@ -206,9 +233,11 @@ func buildSystemConfig(systemConfig configuration.SystemConfig, disks []configur
 			logger.Log.Error("Failed to create the partition map")
 			return
 		}
+		timestamp.StopEvent(nil) // creating delta disk
 	}
 
 	if isOfflineInstall {
+		timestamp.StartEvent("create offline install env", nil)
 		// Create setup chroot
 		additionalExtraMountPoints := []*safechroot.MountPoint{
 			safechroot.NewMountPoint(*assets, assetsMountPoint, "", safechroot.BindMountPointFlags, ""),
@@ -233,13 +262,18 @@ func buildSystemConfig(systemConfig configuration.SystemConfig, disks []configur
 			return
 		}
 
+		timestamp.StopEvent(nil) // create offline install env
+
 		err = setupChroot.Run(func() error {
-			return buildImage(mountPointMap, mountPointToFsTypeMap, mountPointToMountArgsMap, partIDToDevPathMap, partIDToFsTypeMap, mountPointToOverlayMap, packagesToInstall, systemConfig, diskDevPath, isRootFS, encryptedRoot, readOnlyRoot, diffDiskBuild)
+			return buildImage(mountPointMap, mountPointToFsTypeMap, mountPointToMountArgsMap, partIDToDevPathMap, partIDToFsTypeMap, mountPointToOverlayMap, packagesToInstall, systemConfig, diskDevPath, isRootFS, encryptedRoot, readOnlyRoot, diffDiskBuild, imgContentFile)
 		})
 		if err != nil {
 			logger.Log.Error("Failed to build image")
 			return
 		}
+
+		// Extract image package manifest from the 'setuproot' chroot
+		setupChroot.MoveOutFile(installutils.PackageManifestRelativePath, imgContentFile)
 
 		err = cleanupExtraFilesInChroot(setupChroot)
 		if err != nil {
@@ -266,7 +300,7 @@ func buildSystemConfig(systemConfig configuration.SystemConfig, disks []configur
 			}
 		}
 	} else {
-		err = buildImage(mountPointMap, mountPointToFsTypeMap, mountPointToMountArgsMap, partIDToDevPathMap, partIDToFsTypeMap, mountPointToOverlayMap, packagesToInstall, systemConfig, diskDevPath, isRootFS, encryptedRoot, readOnlyRoot, diffDiskBuild)
+		err = buildImage(mountPointMap, mountPointToFsTypeMap, mountPointToMountArgsMap, partIDToDevPathMap, partIDToFsTypeMap, mountPointToOverlayMap, packagesToInstall, systemConfig, diskDevPath, isRootFS, encryptedRoot, readOnlyRoot, diffDiskBuild, imgContentFile)
 		if err != nil {
 			logger.Log.Error("Failed to build image")
 			return
@@ -296,10 +330,10 @@ func setupDiskEncryption(systemConfig *configuration.SystemConfig, encryptedRoot
 
 		// Copy the default keyfile into the image
 		if len(systemConfig.AdditionalFiles) == 0 {
-			systemConfig.AdditionalFiles = make(map[string]string)
+			systemConfig.AdditionalFiles = make(map[string]configuration.FileConfigList)
 		}
 
-		systemConfig.AdditionalFiles[encryptedRoot.HostKeyFile] = diskutils.DefaultKeyFilePath
+		systemConfig.AdditionalFiles[encryptedRoot.HostKeyFile] = configuration.FileConfigList{{Path: diskutils.DefaultKeyFilePath}}
 		logger.Log.Infof("Adding default key file to systemConfig additional files")
 	}
 
@@ -426,8 +460,8 @@ func fixupExtraFilesIntoChroot(installChroot *safechroot.Chroot, config *configu
 		}
 	}
 
-	fixedUpAdditionalFiles := make(map[string]string)
-	for srcFile, dstFile := range config.AdditionalFiles {
+	fixedUpAdditionalFiles := make(map[string]configuration.FileConfigList)
+	for srcFile, dstFileConfigs := range config.AdditionalFiles {
 		newFilePath := filepath.Join(additionalFilesTempDirectory, srcFile)
 
 		fileToCopy := safechroot.FileToCopy{
@@ -435,7 +469,7 @@ func fixupExtraFilesIntoChroot(installChroot *safechroot.Chroot, config *configu
 			Dest: newFilePath,
 		}
 
-		fixedUpAdditionalFiles[newFilePath] = dstFile
+		fixedUpAdditionalFiles[newFilePath] = dstFileConfigs
 		filesToCopy = append(filesToCopy, fileToCopy)
 	}
 	config.AdditionalFiles = fixedUpAdditionalFiles
@@ -489,7 +523,10 @@ func cleanupExtraFilesInChroot(chroot *safechroot.Chroot) (err error) {
 	})
 	return
 }
-func buildImage(mountPointMap, mountPointToFsTypeMap, mountPointToMountArgsMap, partIDToDevPathMap, partIDToFsTypeMap map[string]string, mountPointToOverlayMap map[string]*installutils.Overlay, packagesToInstall []string, systemConfig configuration.SystemConfig, diskDevPath string, isRootFS bool, encryptedRoot diskutils.EncryptedRootDevice, readOnlyRoot diskutils.VerityDevice, diffDiskBuild bool) (err error) {
+
+func buildImage(mountPointMap, mountPointToFsTypeMap, mountPointToMountArgsMap, partIDToDevPathMap, partIDToFsTypeMap map[string]string, mountPointToOverlayMap map[string]*installutils.Overlay, packagesToInstall []string, systemConfig configuration.SystemConfig, diskDevPath string, isRootFS bool, encryptedRoot diskutils.EncryptedRootDevice, readOnlyRoot diskutils.VerityDevice, diffDiskBuild bool, imgContentFile string) (err error) {
+	timestamp.StartEvent("building image", nil)
+	defer timestamp.StopEvent(nil)
 	const (
 		installRoot       = "/installroot"
 		verityWorkingDir  = "verityworkingdir"
@@ -552,6 +589,7 @@ func buildImage(mountPointMap, mountPointToFsTypeMap, mountPointToMountArgsMap, 
 		}
 	}
 
+	timestamp.StartEvent("install chroot packages", nil)
 	for _, setupChrootPackage := range setupChrootPackages {
 		_, err = installutils.TdnfInstall(setupChrootPackage, rootDir)
 		if err != nil {
@@ -559,6 +597,7 @@ func buildImage(mountPointMap, mountPointToFsTypeMap, mountPointToMountArgsMap, 
 			return
 		}
 	}
+	timestamp.StopEvent(nil) // install chroot packages
 
 	// Populate image contents
 	err = installutils.PopulateInstallRoot(installChroot, packagesToInstall, systemConfig, installMap, mountPointToFsTypeMap, mountPointToMountArgsMap, partIDToDevPathMap, partIDToFsTypeMap, isRootFS, encryptedRoot, diffDiskBuild, hidepidEnabled)
@@ -586,6 +625,7 @@ func buildImage(mountPointMap, mountPointToFsTypeMap, mountPointToMountArgsMap, 
 
 		// Snapshot the root filesystem as a read-only verity disk and update the initramfs.
 		if systemConfig.ReadOnlyVerityRoot.Enable {
+			timestamp.StartEvent("configure DM Verity", nil)
 			var initramfsPathList []string
 			err = readOnlyRoot.SwitchDeviceToReadOnly(mountPointMap["/"], mountPointToMountArgsMap["/"])
 			if err != nil {
@@ -602,6 +642,7 @@ func buildImage(mountPointMap, mountPointToFsTypeMap, mountPointToMountArgsMap, 
 				err = fmt.Errorf("failed to include read-only root files in initramfs: %w", err)
 				return
 			}
+			timestamp.StopEvent(nil) // configure DM Verity
 		}
 	}
 
@@ -616,15 +657,18 @@ func buildImage(mountPointMap, mountPointToFsTypeMap, mountPointToMountArgsMap, 
 }
 
 func configureDiskBootloader(systemConfig configuration.SystemConfig, installChroot *safechroot.Chroot, diskDevPath string, installMap map[string]string, encryptedRoot diskutils.EncryptedRootDevice, readOnlyRoot diskutils.VerityDevice) (err error) {
+	timestamp.StartEvent("configuring bootloader", nil)
+	timestamp.StopEvent(nil)
+
 	const rootMountPoint = "/"
 	const bootMountPoint = "/boot"
 
 	var rootDevice string
 
 	// Add bootloader. Prefer a separate boot partition if one exists.
-	bootDevice, ok := installMap[bootMountPoint]
+	bootDevice, isBootPartitionSeparate := installMap[bootMountPoint]
 	bootPrefix := ""
-	if !ok {
+	if !isBootPartitionSeparate {
 		bootDevice = installMap[rootMountPoint]
 		// If we do not have a separate boot partition we will need to add a prefix to all paths used in the configs.
 		bootPrefix = "/boot"
@@ -643,7 +687,7 @@ func configureDiskBootloader(systemConfig configuration.SystemConfig, installChr
 	}
 
 	bootType := systemConfig.BootType
-	err = installutils.InstallBootloader(installChroot, systemConfig.Encryption.Enable, bootType, bootUUID, bootPrefix, diskDevPath)
+	err = installutils.InstallBootloader(installChroot, systemConfig.Encryption.Enable, bootType, bootUUID, bootPrefix, diskDevPath, assetsMountPoint)
 	if err != nil {
 		err = fmt.Errorf("failed to install bootloader: %s", err)
 		return
@@ -679,13 +723,13 @@ func configureDiskBootloader(systemConfig configuration.SystemConfig, installChr
 	}
 
 	// Grub will always use filesystem UUID, never PARTUUID or PARTLABEL
-	err = installutils.InstallGrubCfg(installChroot.RootDir(), rootDevice, bootUUID, bootPrefix, encryptedRoot, systemConfig.KernelCommandLine, readOnlyRoot)
+	err = installutils.InstallGrubCfg(installChroot.RootDir(), rootDevice, bootUUID, bootPrefix, assetsMountPoint, encryptedRoot, systemConfig.KernelCommandLine, readOnlyRoot, isBootPartitionSeparate)
 	if err != nil {
 		err = fmt.Errorf("failed to install main grub config file: %s", err)
 		return
 	}
 
-	err = installutils.InstallGrubEnv(installChroot.RootDir())
+	err = installutils.InstallGrubEnv(installChroot.RootDir(), assetsMountPoint)
 	if err != nil {
 		err = fmt.Errorf("failed to install grubenv file: %s", err)
 		return
