@@ -13,6 +13,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/file"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/logger"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/retry"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/shell"
@@ -30,7 +31,7 @@ func JoinURL(baseURL string, extraPaths ...string) string {
 	return fmt.Sprintf("%s%s%s", baseURL, urlPathSeparator, appendToBase)
 }
 
-// DownloadFile downloads `url` into `dst`. `caCerts` may be nil.
+// DownloadFile downloads `url` into `dst`. `caCerts` may be nil. If there is an error `dst` will be removed.
 func DownloadFile(url, dst string, caCerts *x509.CertPool, tlsCerts []tls.Certificate) (err error) {
 	logger.Log.Debugf("Downloading (%s) -> (%s)", url, dst)
 
@@ -38,7 +39,16 @@ func DownloadFile(url, dst string, caCerts *x509.CertPool, tlsCerts []tls.Certif
 	if err != nil {
 		return
 	}
-	defer dstFile.Close()
+	defer func() {
+		// If there was an error, ensure that the file is removed
+		if err != nil {
+			cleanupErr := file.RemoveFileIfExists(dst)
+			if cleanupErr != nil {
+				logger.Log.Errorf("Failed to remove failed network download file '%s': %s", dst, err)
+			}
+		}
+		dstFile.Close()
+	}()
 
 	tlsConfig := &tls.Config{
 		RootCAs:      caCerts,
