@@ -8,6 +8,6 @@ type RepoManager interface {
 	// CreateRepo will create a package repository at repoDir
 	CreateRepo(repoDir string) error
 
-	// OrganizePackagesByArch will move packages in flatDir into architecture folders under repoDir
-	OrganizePackagesByArch(flatDir, repoDir string) error
+	// ValidateRpmPaths will check if all .rpm files have the correct name as returned from an rpm query
+	ValidateRpmPaths(repoDir string) error
 }
