@@ -9,7 +9,6 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/file"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/logger"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/shell"
 )
@@ -46,52 +45,49 @@ func CreateRepo(repoDir string) (err error) {
 	return
 }
 
-// OrganizePackagesByArch will recursively move RPMs from srcDir into architecture folders under repoDir
-func OrganizePackagesByArch(srcDir, repoDir string) (err error) {
-	const noArch = "noarch"
-
-	logger.Log.Debugf("Organizing RPM packages from (%s) to (%s)", srcDir, repoDir)
-
-	stdout, stderr, err := shell.Execute("uname", "-m")
+// ValidateRpmPaths checks for any rpm filenames in the cache that don't match the expected output according to 'rpm -qp ...'.  It
+// will return an error with all the mismatched pairs if it finds any.
+func ValidateRpmPaths(repoDir string) (err error) {
+	rpmSearch := filepath.Join(repoDir, "*.rpm")
+	rpmFiles, err := filepath.Glob(rpmSearch)
 	if err != nil {
-		logger.Log.Warnf("Could not fetch current architecture from shell: %v", stderr)
 		return
 	}
-	currentArch := strings.TrimSpace(stdout)
 
-	didMovePackages := false
-	pkgArches := []string{noArch, currentArch}
+	// Create a string builder for validation errors
+	validationErrors := []string{}
 
-	for _, arch := range pkgArches {
-		var rpmFiles []string
-
-		err = os.MkdirAll(filepath.Join(repoDir, arch), os.ModePerm)
-		if err != nil {
-			return
-		}
-
-		rpmSearch := filepath.Join(srcDir, fmt.Sprintf("*.%s.rpm", arch))
-		rpmFiles, err = filepath.Glob(rpmSearch)
-		if err != nil {
-			return
-		}
-
-		for _, rpmFile := range rpmFiles {
-			dstFile := filepath.Join(repoDir, arch, filepath.Base(rpmFile))
-			err = file.Move(rpmFile, dstFile)
-			if err != nil {
-				logger.Log.Warnf("Unable to move (%s) to (%s)", rpmFile, dstFile)
-				return
+	for _, rpmFile := range rpmFiles {
+		// rpmFile is the real RPM filename on disk.
+		// use rpm cli to check the rpmFile's reported package name
+		// print a warning if the filename does not match the package name
+		var stdout, stderr string
+		stdout, stderr, err = shell.Execute("rpm", "-qp", rpmFile)
+		if err == nil {
+			calculatedRpmName := strings.Split(stdout, "\n")
+			calculatedRpmFilename := fmt.Sprintf("%s.rpm", calculatedRpmName[0])
+			if calculatedRpmFilename != filepath.Base(rpmFile) {
+				logger.Log.Warnf("!!!!! Detected mismatched filename !!!!!!")
+				logger.Log.Warnf("---- filename   == '%s'", filepath.Base(rpmFile))
+				logger.Log.Warnf("---- calculated == '%s'", calculatedRpmFilename)
+				validationErrors = append(validationErrors, fmt.Sprintf("'%s' != '%s'", filepath.Base(rpmFile), calculatedRpmFilename))
 			}
-		}
-
-		if len(rpmFiles) > 0 {
-			didMovePackages = true
+		} else {
+			err = fmt.Errorf("failed to validate rpm file '%s': '%s'", rpmFile, stderr)
+			return
 		}
 	}
 
-	if !didMovePackages {
-		logger.Log.Debugf("Failed to locate any downloaded packages. All packages are assumed to be locally available.")
+	if len(validationErrors) > 0 {
+		// Now build a single string for the validation errors
+		errorBuilder := strings.Builder{}
+		errorBuilder.WriteString("RPM filename validation errors: ")
+		for _, validationError := range validationErrors {
+			errorBuilder.WriteString(validationError)
+			errorBuilder.WriteString(", ")
+		}
+
+		err = fmt.Errorf(errorBuilder.String())
 	}
 
 	return
