@@ -5,6 +5,7 @@ package rpm
 
 import (
 	"fmt"
+	"path/filepath"
 	"regexp"
 	"runtime"
 	"strings"
@@ -54,10 +55,15 @@ const (
 
 	// MarinerCCacheDefine enables ccache in the Mariner build system
 	MarinerCCacheDefine = "mariner_ccache_enabled"
+
+	// MaxCPUDefine specifies the max number of CPUs to use for parallel build
+	MaxCPUDefine = "_smp_ncpus_max"
 )
 
 const (
-	installedRPMRegexRPMIndex = 1
+	installedRPMRegexRPMIndex        = 1
+	installedRPMRegexArchIndex       = 2
+	installedRPMRegexExpectedMatches = 3
 
 	rpmProgram      = "rpm"
 	rpmSpecProgram  = "rpmspec"
@@ -70,6 +76,10 @@ var (
 		"arm64": "aarch64",
 	}
 
+	// checkSectionRegex is used to determine if a SPEC file has a '%check' section.
+	// It works multi-line strings containing the whole file content, thus the need for the 'm' flag.
+	checkSectionRegex = regexp.MustCompile(`(?m)^\s*%check`)
+
 	// Output from 'rpm' prints installed RPMs in a line with the following format:
 	//
 	//	D: ========== +++ [name]-[version]-[release].[distribution] [architecture]-linux [hex_value]
@@ -77,14 +87,14 @@ var (
 	// Example:
 	//
 	//	D: ========== +++ systemd-devel-239-42.cm2 x86_64-linux 0x0
-	installedRPMRegex = regexp.MustCompile(`^D: =+ \+{3} (\S+).*$`)
+	installedRPMRegex = regexp.MustCompile(`^D: =+ \+{3} (\S+) (\S+)-linux.*$`)
 )
 
 // GetRpmArch converts the GOARCH arch into an RPM arch
 func GetRpmArch(goArch string) (rpmArch string, err error) {
 	rpmArch, ok := goArchToRpmArch[goArch]
 	if !ok {
-		err = fmt.Errorf("Unknown GOARCH detected (%s)", goArch)
+		err = fmt.Errorf("unknown GOARCH detected (%s)", goArch)
 	}
 	return
 }
@@ -99,7 +109,7 @@ func SetMacroDir(newMacroDir string) (origenv []string, err error) {
 	}
 	exists, err := file.DirExists(newMacroDir)
 	if err != nil || exists == false {
-		err = fmt.Errorf("Directory %s does not exist", newMacroDir)
+		err = fmt.Errorf("directory %s does not exist", newMacroDir)
 		return
 	}
 
@@ -107,6 +117,32 @@ func SetMacroDir(newMacroDir string) (origenv []string, err error) {
 	shell.SetEnvironment(env)
 
 	return
+}
+
+// getCommonBuildArgs will generate arguments to pass to 'rpmbuild'.
+func getCommonBuildArgs(outArch, srpmFile string, defines map[string]string) (buildArgs []string, err error) {
+	const (
+		os          = "linux"
+		queryFormat = ""
+		vendor      = "mariner"
+	)
+
+	buildArgs = []string{"--nodeps"}
+
+	// buildArch is the arch of the build machine
+	// outArch is the arch of the machine that will run the resulting binary
+	buildArch, err := GetRpmArch(runtime.GOARCH)
+	if err != nil {
+		return
+	}
+
+	if buildArch != outArch && outArch != "noarch" {
+		tuple := outArch + "-" + vendor + "-" + os
+		logger.Log.Debugf("Applying RPM target tuple (%s)", tuple)
+		buildArgs = append(buildArgs, TargetArgument, tuple)
+	}
+
+	return formatCommandArgs(buildArgs, srpmFile, queryFormat, defines), nil
 }
 
 // sanitizeOutput will take the raw output from an RPM command and split by new line,
@@ -129,7 +165,6 @@ func sanitizeOutput(rawResults string) (sanitizedOutput []string) {
 // formatCommand will generate an RPM command to execute.
 func formatCommandArgs(extraArgs []string, file, queryFormat string, defines map[string]string) (commandArgs []string) {
 	commandArgs = append(commandArgs, extraArgs...)
-	commandArgs = append(commandArgs, file)
 
 	if queryFormat != "" {
 		commandArgs = append(commandArgs, "--qf", queryFormat)
@@ -139,12 +174,21 @@ func formatCommandArgs(extraArgs []string, file, queryFormat string, defines map
 		commandArgs = append(commandArgs, "-D", fmt.Sprintf(`%s %s`, k, v))
 	}
 
+	commandArgs = append(commandArgs, file)
+
 	return
 }
 
 // executeRpmCommand will execute an RPM command and return its output split
 // by new line and whitespace trimmed.
 func executeRpmCommand(program string, args ...string) (results []string, err error) {
+	stdout, err := executeRpmCommandRaw(program, args...)
+
+	return sanitizeOutput(stdout), err
+}
+
+// executeRpmCommandRaw will execute an RPM command and return stdout in form of unmodified strings.
+func executeRpmCommandRaw(program string, args ...string) (stdout string, err error) {
 	stdout, stderr, err := shell.Execute(program, args...)
 	if err != nil {
 		// When dealing with a SPEC/package intended for a different architecture, explicitly set the error message
@@ -157,12 +201,17 @@ func executeRpmCommand(program string, args ...string) (results []string, err er
 		} else {
 			logger.Log.Warn(stderr)
 		}
-
-		return
 	}
 
-	results = sanitizeOutput(stdout)
 	return
+}
+
+// DefaultDefinesWithDist returns a new map of default defines that can be used during RPM queries that also includes
+// the dist tag.
+func DefaultDefinesWithDist(runChecks bool, distTag string) map[string]string {
+	defines := DefaultDefines(runChecks)
+	defines[DistTagDefine] = distTag
+	return defines
 }
 
 // DefaultDefines returns a new map of default defines that can be used during RPM queries.
@@ -191,25 +240,12 @@ func GetInstalledPackages() (result []string, err error) {
 func QuerySPEC(specFile, sourceDir, queryFormat, arch string, defines map[string]string, extraArgs ...string) (result []string, err error) {
 	const queryArg = "-q"
 
-	var allDefines map[string]string
-
 	extraArgs = append(extraArgs, queryArg)
 
 	// Apply --target arch argument
 	extraArgs = append(extraArgs, TargetArgument, arch)
 
-	// To query some SPECs the source directory must be set
-	// since the SPEC file may use `%include` on a source file
-	if sourceDir == "" {
-		allDefines = defines
-	} else {
-		allDefines = make(map[string]string)
-		for k, v := range defines {
-			allDefines[k] = v
-		}
-
-		allDefines[SourceDirDefine] = sourceDir
-	}
+	allDefines := updateSourceDirDefines(defines, sourceDir)
 
 	args := formatCommandArgs(extraArgs, specFile, queryFormat, allDefines)
 	return executeRpmCommand(rpmSpecProgram, args...)
@@ -217,7 +253,7 @@ func QuerySPEC(specFile, sourceDir, queryFormat, arch string, defines map[string
 
 // QuerySPECForBuiltRPMs queries a SPEC file with queryFormat. Returns only the subpackages, which generate a .rpm file.
 func QuerySPECForBuiltRPMs(specFile, sourceDir, arch string, defines map[string]string) (result []string, err error) {
-	const queryFormat = "%{nvra}\n"
+	const queryFormat = "%{nevra}\n"
 
 	return QuerySPEC(specFile, sourceDir, queryFormat, arch, defines, QueryBuiltRPMHeadersArgument)
 }
@@ -232,31 +268,18 @@ func QueryPackage(packageFile, queryFormat string, defines map[string]string, ex
 	return executeRpmCommand(rpmProgram, args...)
 }
 
-// BuildRPMFromSRPM builds an RPM from the given SRPM file
-func BuildRPMFromSRPM(srpmFile, outArch string, defines map[string]string, extraArgs ...string) (err error) {
-	const (
-		queryFormat  = ""
-		squashErrors = true
-		vendor       = "mariner"
-		os           = "linux"
-	)
+// BuildRPMFromSRPM builds an RPM from the given SRPM file but does not run its '%check' section.
+func BuildRPMFromSRPM(srpmFile, outArch string, defines map[string]string) (err error) {
+	const squashErrors = true
 
-	extraArgs = append(extraArgs, "--rebuild", "--nodeps")
-
-	// buildArch is the arch of the build machine
-	// outArch is the arch of the machine that will run the resulting binary
-	buildArch, err := GetRpmArch(runtime.GOARCH)
+	commonBuildArgs, err := getCommonBuildArgs(outArch, srpmFile, defines)
 	if err != nil {
 		return
 	}
 
-	if buildArch != outArch && "noarch" != outArch {
-		tuple := outArch + "-" + vendor + "-" + os
-		logger.Log.Debugf("Applying RPM target tuple (%s)", tuple)
-		extraArgs = append(extraArgs, TargetArgument, tuple)
-	}
+	args := []string{"--nocheck", "--rebuild"}
+	args = append(args, commonBuildArgs...)
 
-	args := formatCommandArgs(extraArgs, srpmFile, queryFormat, defines)
 	return shell.ExecuteLive(squashErrors, rpmBuildProgram, args...)
 }
 
@@ -323,11 +346,6 @@ func QueryRPMProvides(rpmFile string) (provides []string, err error) {
 // ResolveCompetingPackages takes in a list of RPMs and returns only the ones, which would
 // end up being installed after resolving outdated, obsoleted, or conflicting packages.
 func ResolveCompetingPackages(rootDir string, rpmPaths ...string) (resolvedRPMs []string, err error) {
-	const (
-		queryFormat  = ""
-		squashErrors = true
-	)
-
 	args := []string{
 		"-Uvvh",
 		"--replacepkgs",
@@ -349,69 +367,61 @@ func ResolveCompetingPackages(rootDir string, rpmPaths ...string) (resolvedRPMs 
 	uniqueResolvedRPMs := map[string]bool{}
 	for _, line := range splitStdout {
 		matches := installedRPMRegex.FindStringSubmatch(line)
-		if len(matches) != 0 {
-			uniqueResolvedRPMs[matches[installedRPMRegexRPMIndex]] = true
+		if len(matches) == installedRPMRegexExpectedMatches {
+			rpmName := fmt.Sprintf("%s.%s", matches[installedRPMRegexRPMIndex], matches[installedRPMRegexArchIndex])
+			uniqueResolvedRPMs[rpmName] = true
 		}
 	}
 
-	resolvedRPMs = sliceutils.StringsSetToSlice(uniqueResolvedRPMs)
+	resolvedRPMs = sliceutils.SetToSlice(uniqueResolvedRPMs)
 	return
 }
 
 // SpecExclusiveArchIsCompatible verifies the "ExclusiveArch" tag is compatible with the current machine's architecture.
 func SpecExclusiveArchIsCompatible(specfile, sourcedir, arch string, defines map[string]string) (isCompatible bool, err error) {
-	const queryExclusiveArch = "%{ARCH}\n[%{EXCLUSIVEARCH} ]\n"
-
 	const (
-		machineArchField   = iota
-		exclusiveArchField = iota
-		minimumFieldsCount = iota
+		exclusiveArchIndex = 0
+		exclusiveArchQuery = "[%{EXCLUSIVEARCH} ]"
 	)
 
 	// Sanity check that this SPEC is meant to be built for the current machine architecture
-	exclusiveArchList, err := QuerySPEC(specfile, sourcedir, queryExclusiveArch, arch, defines, QueryHeaderArgument)
+	queryOutput, err := QuerySPEC(specfile, sourcedir, exclusiveArchQuery, arch, defines, QueryHeaderArgument)
 	if err != nil {
 		logger.Log.Warnf("Failed to query SPEC (%s), error: %s", specfile, err)
 		return
 	}
 
-	// If the list does not return enough lines then there is no exclusive arch set
-	if len(exclusiveArchList) < minimumFieldsCount {
+	// Empty result means the package is buildable for all architectures.
+	if len(queryOutput) == 0 {
 		isCompatible = true
 		return
 	}
 
-	if strings.Contains(exclusiveArchList[exclusiveArchField], exclusiveArchList[machineArchField]) {
-		isCompatible = true
-		return
-	}
+	isCompatible = strings.Contains(queryOutput[exclusiveArchIndex], arch)
 
 	return
 }
 
 // SpecExcludeArchIsCompatible verifies the "ExcludeArch" tag is compatible with the current machine's architecture.
 func SpecExcludeArchIsCompatible(specfile, sourcedir, arch string, defines map[string]string) (isCompatible bool, err error) {
-	const queryExclusiveArch = "%{ARCH}\n[%{EXCLUDEARCH} ]\n"
-
 	const (
-		machineArchField   = iota
-		excludeArchField   = iota
-		minimumFieldsCount = iota
+		excludeArchIndex = 0
+		excludeArchQuery = "[%{EXCLUDEARCH} ]"
 	)
 
-	excludedArchList, err := QuerySPEC(specfile, sourcedir, queryExclusiveArch, arch, defines, QueryHeaderArgument)
+	queryOutput, err := QuerySPEC(specfile, sourcedir, excludeArchQuery, arch, defines, QueryHeaderArgument)
 	if err != nil {
 		logger.Log.Warnf("Failed to query SPEC (%s), error: %s", specfile, err)
 		return
 	}
 
-	// If the list does not return enough lines then there is no excluded architectures set.
-	if len(excludedArchList) < minimumFieldsCount {
+	// Empty result means the package is buildable for all architectures.
+	if len(queryOutput) == 0 {
 		isCompatible = true
 		return
 	}
 
-	isCompatible = !strings.Contains(excludedArchList[excludeArchField], excludedArchList[machineArchField])
+	isCompatible = !strings.Contains(queryOutput[excludeArchIndex], arch)
 
 	return
 }
@@ -425,6 +435,113 @@ func SpecArchIsCompatible(specfile, sourcedir, arch string, defines map[string]s
 
 	if isCompatible {
 		return SpecExcludeArchIsCompatible(specfile, sourcedir, arch, defines)
+	}
+
+	return
+}
+
+// SpecHasCheckSection verifies if the spec has the '%check' section.
+func SpecHasCheckSection(specFile, sourceDir, arch string, defines map[string]string) (hasCheckSection bool, err error) {
+	const (
+		parseSwitch = "--parse"
+		queryFormat = ""
+	)
+
+	basicArgs := []string{
+		parseSwitch,
+		TargetArgument,
+		arch,
+	}
+	allDefines := updateSourceDirDefines(defines, sourceDir)
+	args := formatCommandArgs(basicArgs, specFile, queryFormat, allDefines)
+
+	stdout, err := executeRpmCommandRaw(rpmSpecProgram, args...)
+
+	return checkSectionRegex.MatchString(stdout), err
+}
+
+// BuildCompatibleSpecsList builds a list of spec files in a directory that are compatible with the build arch. Paths
+// are relative to the 'baseDir' directory. This function should generally be used from inside a chroot to ensure the
+// correct defines are available.
+func BuildCompatibleSpecsList(baseDir string, inputSpecPaths []string, defines map[string]string) (filteredSpecPaths []string, err error) {
+	var specPaths []string
+	if len(inputSpecPaths) > 0 {
+		specPaths = inputSpecPaths
+	} else {
+		specPaths, err = buildAllSpecsList(baseDir)
+		if err != nil {
+			return
+		}
+	}
+
+	return filterCompatibleSpecs(specPaths, defines)
+}
+
+// TestRPMFromSRPM builds an RPM from the given SRPM and runs its '%check' section SRPM file
+// but it does not generate any RPM packages.
+func TestRPMFromSRPM(srpmFile, outArch string, defines map[string]string) (err error) {
+	const squashErrors = true
+
+	commonBuildArgs, err := getCommonBuildArgs(outArch, srpmFile, defines)
+	if err != nil {
+		return
+	}
+
+	args := []string{"-ri"}
+	args = append(args, commonBuildArgs...)
+
+	return shell.ExecuteLive(squashErrors, rpmBuildProgram, args...)
+}
+
+// buildAllSpecsList builds a list of all spec files in the directory. Paths are relative to the base directory.
+func buildAllSpecsList(baseDir string) (specPaths []string, err error) {
+	specFilesGlob := filepath.Join(baseDir, "**", "*.spec")
+
+	specPaths, err = filepath.Glob(specFilesGlob)
+	if err != nil {
+		logger.Log.Errorf("Failed while trying to enumerate all spec files with (%s). Error: %v.", specFilesGlob, err)
+	}
+
+	return
+}
+
+// filterCompatibleSpecs filters a list of spec files in the chroot's SPECs directory that are compatible with the build arch.
+func filterCompatibleSpecs(inputSpecPaths []string, defines map[string]string) (filteredSpecPaths []string, err error) {
+	var specCompatible bool
+
+	buildArch, err := GetRpmArch(runtime.GOARCH)
+	if err != nil {
+		return
+	}
+
+	for _, specFilePath := range inputSpecPaths {
+		specDirPath := filepath.Dir(specFilePath)
+
+		specCompatible, err = SpecArchIsCompatible(specFilePath, specDirPath, buildArch, defines)
+		if err != nil {
+			logger.Log.Errorf("Failed while querrying spec (%s). Error: %v.", specFilePath, err)
+			return
+		}
+
+		if specCompatible {
+			filteredSpecPaths = append(filteredSpecPaths, specFilePath)
+		}
+	}
+
+	return
+}
+
+// updateSourceDirDefines adds the source directory to the defines map if it is not empty.
+// To query some SPECs the source directory must be set
+// since the SPEC file may use `%include` on a source file.
+func updateSourceDirDefines(defines map[string]string, sourceDir string) (updatedDefines map[string]string) {
+	updatedDefines = make(map[string]string)
+	for key, value := range defines {
+		updatedDefines[key] = value
+	}
+
+	if sourceDir != "" {
+		updatedDefines[SourceDirDefine] = sourceDir
 	}
 
 	return
