@@ -9,13 +9,14 @@ import "fmt"
 type BuildAgentConfig struct {
 	Program string
 
-	WorkDir   string
-	WorkerTar string
-	RepoFile  string
-	RpmDir    string
-	SrpmDir   string
-	CacheDir  string
-	CCacheDir string
+	WorkDir      string
+	WorkerTar    string
+	RepoFile     string
+	RpmDir       string
+	ToolchainDir string
+	SrpmDir      string
+	CacheDir     string
+	CCacheDir    string
 
 	DistTag              string
 	DistroReleaseVersion string
@@ -23,8 +24,8 @@ type BuildAgentConfig struct {
 	RpmmacrosFile        string
 
 	NoCleanup bool
-	RunCheck  bool
 	UseCcache bool
+	MaxCpu    string
 
 	LogDir   string
 	LogLevel string
@@ -38,9 +39,10 @@ type BuildAgent interface {
 	// BuildPackage builds a given file and returns the output files or error.
 	// - inputFile is the SRPM to build.
 	// - logName is the file name to save the package build log to.
-	// - outArch is the machine architecture where the output binary will run
+	// - outArch is the target architecture to build for.
+	// - runCheck is true if the package should run the "%check" section during the build
 	// - dependencies is a list of dependencies that need to be installed before building.
-	BuildPackage(inputFile, logName, outArch string, dependencies []string) ([]string, string, error)
+	BuildPackage(inputFile, logName, outArch string, runCheck bool, dependencies []string) ([]string, string, error)
 
 	// Config returns a copy of the agent's configuration.
 	Config() BuildAgentConfig
