@@ -8,7 +8,10 @@ import (
 
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/logger"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/pkggraph"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/sliceutils"
+	"github.com/sirupsen/logrus"
 	"gonum.org/v1/gonum/graph"
+	"gonum.org/v1/gonum/graph/path"
 	"gonum.org/v1/gonum/graph/traverse"
 )
 
@@ -40,6 +43,20 @@ func CanSubGraph(pkgGraph *pkggraph.PkgGraph, node *pkggraph.PkgNode, useCachedI
 
 		// This node is not yet solvable
 		logger.Log.Warnf("Could not subgraph due to node: %v", pkgNode)
+
+		// If we are in trace mode, print the path from the root node to the unsolvable node
+		if logger.Log.IsLevelEnabled(logrus.TraceLevel) {
+			paths := path.YenKShortestPaths(pkgGraph, 1, node, pkgNode)
+			if len(paths) == 0 {
+				logger.Log.Warnf("Could not find path between %v and %v with YenKShortestPaths()", node, pkgNode)
+			} else {
+				logger.Log.Tracef("Path between %v and %v:", node, pkgNode)
+				for _, n := range paths[0] {
+					logger.Log.Tracef("  %v", n.(*pkggraph.PkgNode))
+				}
+			}
+		}
+
 		foundUnsolvableNode = true
 		return
 	})
@@ -72,6 +89,9 @@ func LeafNodes(pkgGraph *pkggraph.PkgGraph, graphMutex *sync.RWMutex, goalNode *
 			return
 		}
 
+		// Implicit nodes will only be considered valid leaf nodes as a last resort (aka when useCachedImplicit is set).
+		// Ideally we will wait for the actual provider of the implicit node to be built and convert the implicit node to
+		// a normal node via InjectMissingImplicitProvides().
 		if !useCachedImplicit && pkgNode.Implicit && pkgNode.State == pkggraph.StateCached {
 			logger.Log.Debugf("Skipping cached implicit provide leaf node: %v", pkgNode)
 			return
@@ -100,36 +120,25 @@ func FindUnblockedNodesFromResult(res *BuildResult, pkgGraph *pkggraph.PkgGraph,
 	// Since all the ancillary nodes are marked as available already, there may be duplicate nodes returned by the below loop.
 	// e.g. If a meta node requires two build nodes for the same SPEC, then that meta node will be reported twice.
 	// Filter the nodes to ensure no duplicates.
-	var unfilteredUnblockedNodes []*pkggraph.PkgNode
 	unblockedNodesMap := make(map[*pkggraph.PkgNode]bool)
 	for _, node := range res.AncillaryNodes {
-		unfilteredUnblockedNodes = append(unfilteredUnblockedNodes, findUnblockedNodesFromNode(pkgGraph, buildState, node)...)
+		findUnblockedNodesFromNode(pkgGraph, buildState, node, unblockedNodesMap)
 	}
 
-	for _, node := range unfilteredUnblockedNodes {
-		_, found := unblockedNodesMap[node]
-		if !found {
-			unblockedNodesMap[node] = true
-			unblockedNodes = append(unblockedNodes, node)
-		}
-	}
-
-	return
+	return sliceutils.SetToSlice(unblockedNodesMap)
 }
 
 // findUnblockedNodesFromNode takes a built node and returns a list of nodes that are now unblocked by it.
-func findUnblockedNodesFromNode(pkgGraph *pkggraph.PkgGraph, buildState *GraphBuildState, builtNode *pkggraph.PkgNode) (unblockedNodes []*pkggraph.PkgNode) {
+func findUnblockedNodesFromNode(pkgGraph *pkggraph.PkgGraph, buildState *GraphBuildState, builtNode *pkggraph.PkgNode, unblockedNodes map[*pkggraph.PkgNode]bool) {
 	dependents := pkgGraph.To(builtNode.ID())
 
 	for dependents.Next() {
 		dependent := dependents.Node().(*pkggraph.PkgNode)
 
 		if isNodeUnblocked(pkgGraph, buildState, dependent) {
-			unblockedNodes = append(unblockedNodes, dependent)
+			unblockedNodes[dependent] = true
 		}
 	}
-
-	return
 }
 
 // isNodeUnblocked returns true if all nodes required to build `node` are UpToDate and do not need to be built.
