@@ -4,6 +4,8 @@
 package schedulerutils
 
 import (
+	"fmt"
+	"math"
 	"path/filepath"
 	"sort"
 
@@ -16,12 +18,20 @@ import (
 type nodeState struct {
 	available bool
 	cached    bool
+	usedDelta bool
+	freshness uint
 }
+
+const (
+	// The highest possible freshness value, used to force unbounded cascading rebuilds.
+	NodeFreshnessAbsoluteMax uint64 = math.MaxUint64
+)
 
 // GraphBuildState represents the build state of a graph.
 type GraphBuildState struct {
 	activeBuilds     map[int64]*BuildRequest
 	nodeToState      map[*pkggraph.PkgNode]*nodeState
+	maxFreshness     uint
 	failures         []*BuildResult
 	reservedFiles    map[string]bool
 	conflictingRPMs  map[string]bool
@@ -29,18 +39,20 @@ type GraphBuildState struct {
 }
 
 // NewGraphBuildState returns a new GraphBuildState.
-// - reservedFiles is a list of reserved files which should NOT be rebuilt. Any files that ARE rebuilt will be recorded.
-func NewGraphBuildState(reservedFiles []string) (g *GraphBuildState) {
-	filesMap := make(map[string]bool)
-	for _, file := range reservedFiles {
-		filesMap[file] = true
-	}
+//   - reservedFiles is a list of reserved files which should NOT be rebuilt. Any files that ARE rebuilt will be recorded.
+//   - maxFreshness is how fresh a newly rebuilt node is. Each dependant node will have a freshness of 'n-1', etc. until
+//     '0' where the subsequent nodes will no longer be rebuilt. 'maxFreshness < 0' will cause unbounded cascading rebuilds,
+//     while 'maxFreshness = 0' will cause no cascading rebuilds.
+func NewGraphBuildState(reservedFiles []string, maxFreshness uint) (g *GraphBuildState) {
+	filesMap := sliceutils.SliceToSet[string](reservedFiles)
+
 	return &GraphBuildState{
 		activeBuilds:     make(map[int64]*BuildRequest),
 		nodeToState:      make(map[*pkggraph.PkgNode]*nodeState),
 		reservedFiles:    filesMap,
 		conflictingRPMs:  make(map[string]bool),
 		conflictingSRPMs: make(map[string]bool),
+		maxFreshness:     maxFreshness,
 	}
 }
 
@@ -67,16 +79,67 @@ func (g *GraphBuildState) IsNodeCached(node *pkggraph.PkgNode) bool {
 	return state != nil && state.cached
 }
 
+// GetMaxFreshness returns the maximum freshness a node can have. (ie if a package is directly rebuilt due to user
+// request, or missing files, it will have this freshness. Each dependant node will have a freshness of 'n-1', etc.
+func (g *GraphBuildState) GetMaxFreshness() uint {
+	return g.maxFreshness
+}
+
+// GetFreshnessOfNode returns the freshness of a node.
+func (g *GraphBuildState) GetFreshnessOfNode(node *pkggraph.PkgNode) uint {
+	return g.nodeToState[node].freshness
+}
+
+// IsNodeDelta returns true if the requested node was pre-downloaded as a delta package.
+func (g *GraphBuildState) IsNodeDelta(node *pkggraph.PkgNode) bool {
+	state := g.nodeToState[node]
+	return state != nil && state.usedDelta
+}
+
+// IsSRPMCached returns true if the requested SRPM build used a cache.
+func (g *GraphBuildState) IsSRPMCached(srpmFileName string) bool {
+	for node, state := range g.nodeToState {
+		if node.Type == pkggraph.TypeLocalBuild && node.SRPMFileName() == srpmFileName {
+			return state.cached
+		}
+	}
+
+	return false
+}
+
 // ActiveBuilds returns a map of Node IDs to BuildRequests that represents all outstanding builds.
 func (g *GraphBuildState) ActiveBuilds() map[int64]*BuildRequest {
 	return g.activeBuilds
 }
 
+// ActiveBuildFromSRPM returns a build request for the queried SRPM file
+// or nil if the SRPM is not among the active builds.
+func (g *GraphBuildState) ActiveBuildFromSRPM(srpmFileName string) *BuildRequest {
+	for _, buildRequest := range g.activeBuilds {
+		if buildRequest.Node.Type == pkggraph.TypeLocalBuild && buildRequest.Node.SRPMFileName() == srpmFileName {
+			return buildRequest
+		}
+	}
+
+	return nil
+}
+
 // ActiveSRPMs returns a list of all SRPMs, which are currently being built.
 func (g *GraphBuildState) ActiveSRPMs() (builtSRPMs []string) {
 	for _, buildRequest := range g.activeBuilds {
-		if buildRequest.Node.Type == pkggraph.TypeBuild {
+		if buildRequest.Node.Type == pkggraph.TypeLocalBuild {
 			builtSRPMs = append(builtSRPMs, buildRequest.Node.SRPMFileName())
+		}
+	}
+
+	return
+}
+
+// ActiveTests returns a list of all tests, which are currently being run.
+func (g *GraphBuildState) ActiveTests() (testedSRPMs []string) {
+	for _, testRequest := range g.activeBuilds {
+		if testRequest.Node.Type == pkggraph.TypeTest {
+			testedSRPMs = append(testedSRPMs, testRequest.Node.SRPMFileName())
 		}
 	}
 
@@ -91,7 +154,7 @@ func (g *GraphBuildState) BuildFailures() []*BuildResult {
 // ConflictingRPMs will return a list of *.rpm files which should not have been rebuilt.
 // This list is based on the manifest of pre-built toolchain rpms.
 func (g *GraphBuildState) ConflictingRPMs() (rpms []string) {
-	rpms = sliceutils.StringsSetToSlice(g.conflictingRPMs)
+	rpms = sliceutils.SetToSlice(g.conflictingRPMs)
 	sort.Strings(rpms)
 
 	return rpms
@@ -100,7 +163,7 @@ func (g *GraphBuildState) ConflictingRPMs() (rpms []string) {
 // ConflictingSRPMs will return a list of *.src.rpm files which created rpms that should not have been rebuilt.
 // This list is based on the manifest of pre-built toolchain rpms.
 func (g *GraphBuildState) ConflictingSRPMs() (srpms []string) {
-	srpms = sliceutils.StringsSetToSlice(g.conflictingSRPMs)
+	srpms = sliceutils.SetToSlice(g.conflictingSRPMs)
 	sort.Strings(srpms)
 
 	return srpms
@@ -126,8 +189,7 @@ func (g *GraphBuildState) isConflictWithToolchain(fileToCheck string) (hadConfli
 // RecordBuildResult records a build result in the graph build state.
 // - It will record the result as a failure if applicable.
 // - It will record all ancillary nodes of the result.
-func (g *GraphBuildState) RecordBuildResult(res *BuildResult) {
-
+func (g *GraphBuildState) RecordBuildResult(res *BuildResult, allowToolchainRebuilds bool) (err error) {
 	logger.Log.Debugf("Recording build result: %s", res.Node.FriendlyName())
 
 	delete(g.activeBuilds, res.Node.ID())
@@ -136,16 +198,27 @@ func (g *GraphBuildState) RecordBuildResult(res *BuildResult) {
 		g.failures = append(g.failures, res)
 	}
 
+	// 'NodeFreshnessRebuildRequired' is a special value that indicates that the node was rebuilt due to  missing files
+	// (user requested rebuilds are already at the max freshness). In this case, we want to reset the freshness to the
+	// max, so that subsequent dependant nodes will be rebuilt. Also ensure that the freshness is not greater than the max.
+	freshness := res.Freshness
+	if freshness > g.GetMaxFreshness() {
+		err = fmt.Errorf("unexpected freshness value of %d for node '%s'. Should be: (<freshness> <= %d)", freshness, res.Node.FriendlyName(), g.GetMaxFreshness())
+		return
+	}
+
 	state := &nodeState{
 		available: res.Err == nil,
 		cached:    res.UsedCache,
+		usedDelta: res.WasDelta,
+		freshness: freshness,
 	}
 
 	for _, node := range res.AncillaryNodes {
 		g.nodeToState[node] = state
 	}
 
-	if !res.Skipped && !res.UsedCache {
+	if !allowToolchainRebuilds && !res.Ignored && !res.UsedCache {
 		for _, file := range res.BuiltFiles {
 			if g.isConflictWithToolchain(file) {
 				g.conflictingRPMs[filepath.Base(file)] = true
@@ -153,8 +226,7 @@ func (g *GraphBuildState) RecordBuildResult(res *BuildResult) {
 			}
 		}
 	} else {
-		logger.Log.Debugf("skipping checking conflicts since this is not a built node (%v)", res.Node)
+		logger.Log.Tracef("Skipping checking toolchain conflicts since this is either not a built node (%v) or the ALLOW_TOOLCHAIN_REBUILDS flag was set to 'y'.", res.Node)
 	}
-
 	return
 }
