@@ -20,16 +20,40 @@ import (
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/exe"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/file"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/logger"
+	packagelist "github.com/microsoft/CBL-Mariner/toolkit/tools/internal/packlist"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/pkgjson"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/rpm"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/safechroot"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/timestamp"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/pkg/profile"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/scheduler/schedulerutils"
 
 	"github.com/jinzhu/copier"
 	"gopkg.in/alecthomas/kingpin.v2"
 )
 
 const (
-	defaultWorkerCount = "10"
+	// Spaces added on purpose to simplify substring matching.
+	andCondition  = " and "
+	ifCondition   = " if "
+	orCondition   = " or "
+	withCondition = " with "
+)
+
+var (
+	supportedBooleanConditions = []string{
+		andCondition,
+		ifCondition,
+		orCondition,
+		withCondition,
+	}
+
+	// Spaces added on purpose to simplify substring matching.
+	unsupportedBooleanConditions = []string{
+		" else ",
+		" unless ",
+		" without ",
+	}
 )
 
 // parseResult holds the worker results from parsing a SPEC file.
@@ -38,20 +62,29 @@ type parseResult struct {
 	err      error
 }
 
+const (
+	defaultWorkerCount = "100"
+)
+
 var (
-	app        = kingpin.New("specreader", "A tool to parse spec dependencies into JSON")
-	specsDir   = exe.InputDirFlag(app, "Directory to scan for SPECS")
-	output     = exe.OutputFlag(app, "Output file to export the JSON")
-	workers    = app.Flag("workers", "Number of concurrent goroutines to parse with").Default(defaultWorkerCount).Int()
-	buildDir   = app.Flag("build-dir", "Directory to store temporary files while parsing.").String()
-	srpmsDir   = app.Flag("srpm-dir", "Directory containing SRPMs.").Required().ExistingDir()
-	rpmsDir    = app.Flag("rpm-dir", "Directory containing built RPMs.").Required().ExistingDir()
-	distTag    = app.Flag("dist-tag", "The distribution tag the SPEC will be built with.").Required().String()
-	workerTar  = app.Flag("worker-tar", "Full path to worker_chroot.tar.gz.  If this argument is empty, specs will be parsed in the host environment.").ExistingFile()
-	targetArch = app.Flag("target-arch", "The architecture of the machine the RPM binaries run on").String()
-	runCheck   = app.Flag("run-check", "Whether or not to run the spec file's check section during package build.").Bool()
-	logFile    = exe.LogFileFlag(app)
-	logLevel   = exe.LogLevelFlag(app)
+	app                     = kingpin.New("specreader", "A tool to parse spec dependencies into JSON")
+	specsDir                = exe.InputDirFlag(app, "Directory to scan for SPECS")
+	specListFile            = app.Flag("spec-list", "Path to a list of SPECs to parse. If empty will parse all SPECs.").ExistingFile()
+	output                  = exe.OutputFlag(app, "Output file to export the JSON")
+	workers                 = app.Flag("workers", "Number of concurrent goroutines to parse with").Default(defaultWorkerCount).Int()
+	buildDir                = app.Flag("build-dir", "Directory to store temporary files while parsing.").String()
+	srpmsDir                = app.Flag("srpm-dir", "Directory containing SRPMs.").Required().ExistingDir()
+	rpmsDir                 = app.Flag("rpm-dir", "Directory containing built RPMs.").Required().ExistingDir()
+	toolchainManifest       = app.Flag("toolchain-manifest", "Path to a list of RPMs which are created by the toolchain. Will mark RPMs from this list as prebuilt.").ExistingFile()
+	existingToolchainRpmDir = app.Flag("toolchain-rpms-dir", "Directory that contains already built toolchain RPMs. Should contain top level directories for architecture.").Required().ExistingDir()
+	distTag                 = app.Flag("dist-tag", "The distribution tag the SPEC will be built with.").Required().String()
+	workerTar               = app.Flag("worker-tar", "Full path to worker_chroot.tar.gz.  If this argument is empty, specs will be parsed in the host environment.").ExistingFile()
+	targetArch              = app.Flag("target-arch", "The architecture of the machine the RPM binaries run on").String()
+	runCheck                = app.Flag("run-check", "Whether or not to run the spec file's check section during package build.").Bool()
+	logFile                 = exe.LogFileFlag(app)
+	logLevel                = exe.LogLevelFlag(app)
+	profFlags               = exe.SetupProfileFlags(app)
+	timestampFile           = app.Flag("timestamp-file", "File that stores timestamps for this program.").String()
 )
 
 func main() {
@@ -59,17 +92,38 @@ func main() {
 	kingpin.MustParse(app.Parse(os.Args[1:]))
 	logger.InitBestEffort(*logFile, *logLevel)
 
+	prof, err := profile.StartProfiling(profFlags)
+	if err != nil {
+		logger.Log.Warnf("Could not start profiling: %s", err)
+	}
+	defer prof.StopProfiler()
+
+	timestamp.BeginTiming("specreader", *timestampFile)
+	defer timestamp.CompleteTiming()
+
 	if *workers <= 0 {
 		logger.Log.Panicf("Value in --workers must be greater than zero. Found %d", *workers)
 	}
 
-	err := parseSPECsWrapper(*buildDir, *specsDir, *rpmsDir, *srpmsDir, *distTag, *output, *workerTar, *workers, *runCheck)
+	toolchainRPMs, err := schedulerutils.ReadReservedFilesList(*toolchainManifest)
+	logger.PanicOnError(err, "Unable to read toolchain manifest file '%s': %s", *toolchainManifest, err)
+
+	// A parse list may be provided, if so only parse this subset.
+	// If none is provided, parse all specs.
+	specListSet, err := packagelist.ParsePackageListFile(*specListFile)
+	logger.PanicOnError(err)
+
+	// Convert specsDir to an absolute path
+	specsAbsDir, err := filepath.Abs(*specsDir)
+	logger.PanicOnError(err, "Unable to get absolute path for specs directory '%s': %s", *specsDir, err)
+
+	err = parseSPECsWrapper(*buildDir, specsAbsDir, *rpmsDir, *srpmsDir, *existingToolchainRpmDir, *distTag, *output, *workerTar, specListSet, toolchainRPMs, *workers, *runCheck)
 	logger.PanicOnError(err)
 }
 
 // parseSPECsWrapper wraps parseSPECs to conditionally run it inside a chroot.
 // If workerTar is non-empty, parsing will occur inside a chroot, otherwise it will run on the host system.
-func parseSPECsWrapper(buildDir, specsDir, rpmsDir, srpmsDir, distTag, outputFile, workerTar string, workers int, runCheck bool) (err error) {
+func parseSPECsWrapper(buildDir, specsDir, rpmsDir, srpmsDir, toolchainDir, distTag, outputFile, workerTar string, specListSet map[string]bool, toolchainRPMs []string, workers int, runCheck bool) (err error) {
 	var (
 		chroot      *safechroot.Chroot
 		packageRepo *pkgjson.PackageRepo
@@ -93,15 +147,15 @@ func parseSPECsWrapper(buildDir, specsDir, rpmsDir, srpmsDir, distTag, outputFil
 		var parseError error
 
 		if *targetArch == "" {
-			packageRepo, parseError = parseSPECs(specsDir, rpmsDir, srpmsDir, distTag, buildArch, workers, runCheck)
+			packageRepo, parseError = parseSPECs(specsDir, rpmsDir, srpmsDir, toolchainDir, distTag, buildArch, specListSet, toolchainRPMs, workers, runCheck)
 			if parseError != nil {
-				err := fmt.Errorf("Failed to parse native specs (%w)", parseError)
+				err := fmt.Errorf("failed to parse native specs (%w)", parseError)
 				return err
 			}
 		} else {
-			packageRepo, parseError = parseSPECs(specsDir, rpmsDir, srpmsDir, distTag, *targetArch, workers, runCheck)
+			packageRepo, parseError = parseSPECs(specsDir, rpmsDir, srpmsDir, toolchainDir, distTag, *targetArch, specListSet, toolchainRPMs, workers, runCheck)
 			if parseError != nil {
-				err := fmt.Errorf("Failed to parse cross specs (%w)", parseError)
+				err := fmt.Errorf("failed to parse cross specs (%w)", parseError)
 				return err
 			}
 		}
@@ -181,8 +235,41 @@ func createChroot(workerTar, buildDir, specsDir, srpmsDir string) (chroot *safec
 	return
 }
 
+func findSpecFiles(specsDir string, specListSet map[string]bool) (specFiles []string, err error) {
+	// Find the filepath for each spec in the SPECS directory.
+	if len(specListSet) == 0 {
+		specSearch, err := filepath.Abs(filepath.Join(specsDir, "**/*.spec"))
+		if err != nil {
+			err = fmt.Errorf("invalid spec dir: '%s'. Error:\n%w", specsDir, err)
+			return nil, err
+		}
+		specFiles, err = filepath.Glob(specSearch)
+		if err != nil {
+			err = fmt.Errorf("failed to find *.spec files. Check that '%s' is the correct directory. Error:\n%w", specsDir, err)
+			return nil, err
+		}
+	} else {
+		for specName := range specListSet {
+			specSearch := filepath.Join(specsDir, fmt.Sprintf("**/%s.spec", specName))
+			matchingSpecFiles, err := filepath.Glob(specSearch)
+
+			// If a SPEC is in the parse list, it should be parsed.
+			if err != nil {
+				err = fmt.Errorf("spec search failed on '%s'. Error:\n%w", specSearch, err)
+				return nil, err
+			}
+			if len(matchingSpecFiles) != 1 {
+				err = fmt.Errorf("unexpected number of matches '%d' for spec file '%s'", len(matchingSpecFiles), specName)
+				return nil, err
+			}
+			specFiles = append(specFiles, matchingSpecFiles[0])
+		}
+	}
+	return
+}
+
 // parseSPECs will parse all specs in specsDir and return a summary of the SPECs.
-func parseSPECs(specsDir, rpmsDir, srpmsDir, distTag, arch string, workers int, runCheck bool) (packageRepo *pkgjson.PackageRepo, err error) {
+func parseSPECs(specsDir, rpmsDir, srpmsDir, toolchainDir, distTag, arch string, specListSet map[string]bool, toolchainRPMs []string, workers int, runCheck bool) (packageRepo *pkgjson.PackageRepo, err error) {
 	var (
 		packageList []*pkgjson.Package
 		wg          sync.WaitGroup
@@ -191,15 +278,14 @@ func parseSPECs(specsDir, rpmsDir, srpmsDir, distTag, arch string, workers int, 
 
 	packageRepo = &pkgjson.PackageRepo{}
 
-	// Find the filepath for each spec in the SPECS directory.
-	specSearch, err := filepath.Abs(filepath.Join(specsDir, "**/*.spec"))
-	if err == nil {
-		specFiles, err = filepath.Glob(specSearch)
-	}
+	specFiles, err = findSpecFiles(specsDir, specListSet)
 	if err != nil {
 		logger.Log.Errorf("Failed to find *.spec files. Check that %s is the correct directory. Error: %v", specsDir, err)
 		return
 	}
+
+	tsRoot, _ := timestamp.StartEvent("parse specs", nil)
+	defer timestamp.StopEvent(nil)
 
 	results := make(chan *parseResult, len(specFiles))
 	requests := make(chan string, len(specFiles))
@@ -208,7 +294,7 @@ func parseSPECs(specsDir, rpmsDir, srpmsDir, distTag, arch string, workers int, 
 	// Start the workers now so they begin working as soon as a new job is buffered.
 	for i := 0; i < workers; i++ {
 		wg.Add(1)
-		go readSpecWorker(requests, results, cancel, &wg, distTag, rpmsDir, srpmsDir, runCheck, arch)
+		go readSpecWorker(requests, results, cancel, &wg, distTag, rpmsDir, srpmsDir, toolchainDir, toolchainRPMs, runCheck, arch, tsRoot)
 	}
 
 	for _, specFile := range specFiles {
@@ -265,22 +351,22 @@ func sortPackages(packageRepo *pkgjson.PackageRepo) {
 	}
 }
 
-// readspec is a goroutine that takes a full filepath to a spec file and scrapes it into the Specdef structure
+// readSpecWorker is a goroutine that takes a full filepath to a spec file and scrapes it into the Specdef structure
 // Concurrency is limited by the size of the semaphore channel passed in. Too many goroutines at once can deplete
-// available filehandles.
-func readSpecWorker(requests <-chan string, results chan<- *parseResult, cancel <-chan struct{}, wg *sync.WaitGroup, distTag, rpmsDir, srpmsDir string, runCheck bool, arch string) {
+// available file handles.
+func readSpecWorker(requests <-chan string, results chan<- *parseResult, cancel <-chan struct{}, wg *sync.WaitGroup, distTag, rpmsDir, srpmsDir, toolchainDir string, toolchainRPMs []string, runCheck bool, arch string, tsRoot *timestamp.TimeStamp) {
 	const (
-		emptyQueryFormat      = ``
 		querySrpm             = `%{NAME}-%{VERSION}-%{RELEASE}.src.rpm`
 		queryProvidedPackages = `rpm %{ARCH}/%{nvra}.rpm\n[provides %{PROVIDENEVRS}\n][requires %{REQUIRENEVRS}\n][arch %{ARCH}\n]`
 	)
 
 	defer wg.Done()
 
-	defines := rpm.DefaultDefines(runCheck)
-	defines[rpm.DistTagDefine] = distTag
+	noCheckDefines := rpm.DefaultDefinesWithDist(false, distTag)
+	checkDefines := rpm.DefaultDefinesWithDist(true, distTag)
 
-	for specfile := range requests {
+	var ts *timestamp.TimeStamp = nil
+	for specFile := range requests {
 		select {
 		case <-cancel:
 			logger.Log.Debug("Cancellation signal received")
@@ -288,80 +374,90 @@ func readSpecWorker(requests <-chan string, results chan<- *parseResult, cancel 
 		default:
 		}
 
-		result := &parseResult{}
+		// Many code paths hit 'continue', finish timing those here.
+		if ts != nil {
+			timestamp.StopEvent(ts)
+			ts = nil
+		}
+		ts, _ = timestamp.StartEvent(filepath.Base(specFile), tsRoot)
 
 		providerList := []*pkgjson.Package{}
-		buildRequiresList := []*pkgjson.PackageVer{}
-		sourcedir := filepath.Dir(specfile)
+		sourceDir := filepath.Dir(specFile)
+		testBuildRequiresList := []*pkgjson.PackageVer{}
 
 		// Find the SRPM associated with the SPEC.
-		srpmResults, err := rpm.QuerySPEC(specfile, sourcedir, querySrpm, arch, defines, rpm.QueryHeaderArgument)
+		srpmResults, err := rpm.QuerySPEC(specFile, sourceDir, querySrpm, arch, noCheckDefines, rpm.QueryHeaderArgument)
 		if err != nil {
-			result.err = err
-			results <- result
+			sendEmptyResult(results, err)
 			continue
 		}
 
 		srpmPath := filepath.Join(srpmsDir, srpmResults[0])
 
-		isCompatible, err := rpm.SpecArchIsCompatible(specfile, sourcedir, arch, defines)
+		isCompatible, err := rpm.SpecArchIsCompatible(specFile, sourceDir, arch, noCheckDefines)
 		if err != nil {
-			result.err = err
-			results <- result
+			sendEmptyResult(results, err)
 			continue
 		}
 
 		if !isCompatible {
-			logger.Log.Debugf(`Skipping (%s) since it cannot be built on current architecture.`, specfile)
-			results <- result
+			logger.Log.Debugf(`Skipping (%s) since it cannot be built on current architecture.`, specFile)
+			sendEmptyResult(results, err)
 			continue
 		}
 
 		// Find every package that the spec provides
-		queryResults, err := rpm.QuerySPEC(specfile, sourcedir, queryProvidedPackages, arch, defines, rpm.QueryBuiltRPMHeadersArgument)
-		if err == nil && len(queryResults) != 0 {
-			providerList, err = parseProvides(rpmsDir, srpmPath, queryResults)
+		queryResults, err := rpm.QuerySPEC(specFile, sourceDir, queryProvidedPackages, arch, noCheckDefines, rpm.QueryBuiltRPMHeadersArgument)
+		if err != nil {
+			sendEmptyResult(results, err)
+			continue
+		}
+
+		if len(queryResults) != 0 {
+			providerList, err = parseProvides(rpmsDir, toolchainDir, toolchainRPMs, srpmPath, queryResults)
 			if err != nil {
-				result.err = err
-				results <- result
+				sendEmptyResult(results, err)
 				continue
 			}
 		}
 
 		// Query the BuildRequires fields from this spec and turn them into an array of PackageVersions
-		queryResults, err = rpm.QuerySPEC(specfile, sourcedir, emptyQueryFormat, arch, defines, rpm.BuildRequiresArgument)
-		if err == nil && len(queryResults) != 0 {
-			buildRequiresList, err = parsePackageVersionList(queryResults)
+		buildRequiresList, err := readBuildRequires(specFile, sourceDir, arch, noCheckDefines)
+		if err != nil {
+			sendEmptyResult(results, err)
+			continue
+		}
+
+		specHasCheckSection, err := rpm.SpecHasCheckSection(specFile, sourceDir, arch, checkDefines)
+		if err != nil {
+			sendEmptyResult(results, err)
+			continue
+		}
+
+		readTestDependencies := runCheck && specHasCheckSection
+		if readTestDependencies {
+			// Query the test BuildRequires fields from this spec and turn them into an array of PackageVersions
+			testBuildRequiresList, err = readBuildRequires(specFile, sourceDir, arch, checkDefines)
 			if err != nil {
-				result.err = err
-				results <- result
+				sendEmptyResult(results, err)
 				continue
 			}
 		}
 
 		// Every package provided by a spec will have the same BuildRequires and SrpmPath
-		for i := range providerList {
-			providerList[i].SpecPath = specfile
-			providerList[i].SourceDir = sourcedir
-			providerList[i].Requires, err = condensePackageVersionArray(providerList[i].Requires, specfile)
-			if err != nil {
-				break
-			}
-
-			providerList[i].BuildRequires, err = condensePackageVersionArray(buildRequiresList, specfile)
-			if err != nil {
-				break
-			}
-		}
-
-		if err != nil {
-			result.err = err
-		} else {
-			result.packages = providerList
+		for _, provider := range providerList {
+			provider.BuildRequires = buildRequiresList
+			provider.SourceDir = sourceDir
+			provider.SpecPath = specFile
+			provider.TestRequires = testBuildRequiresList
+			provider.RunTests = readTestDependencies
 		}
 
 		// Submit the result to the main thread, the deferred function will clear the semaphore.
-		results <- result
+		results <- &parseResult{packages: providerList}
+	}
+	if ts != nil {
+		timestamp.StopEvent(ts)
 	}
 }
 
@@ -373,7 +469,7 @@ func readSpecWorker(requests <-chan string, results chan<- *parseResult, cancel 
 // Require: requiresb
 // Arch: noarch
 // The return is an array of Package structures, one for each Provides in the spec (implicit and explicit).
-func parseProvides(rpmsDir, srpmPath string, list []string) (providerlist []*pkgjson.Package, err error) {
+func parseProvides(rpmsDir, toolchainDir string, toolchainRPMs []string, srpmPath string, list []string) (providerlist []*pkgjson.Package, err error) {
 	var (
 		reqlist      []*pkgjson.PackageVer
 		packagearch  string
@@ -410,6 +506,7 @@ func parseProvides(rpmsDir, srpmPath string, list []string) (providerlist []*pkg
 
 		if listEntry[tag] == "rpm" {
 			logger.Log.Trace("rpm ", listEntry[value])
+			// We will decide if this is a toolchain package later, and update the path if so.
 			rpmPath = filepath.Join(rpmsDir, listEntry[value])
 		} else if listEntry[tag] == "provides" {
 			logger.Log.Trace("provides ", listEntry[value])
@@ -442,6 +539,17 @@ func parseProvides(rpmsDir, srpmPath string, list []string) (providerlist []*pkg
 				return
 			}
 
+			reqlist, err = dedupPackageVersionArray(reqlist)
+			if err != nil {
+				err = fmt.Errorf("failed to dedup run-time PackageVer array for SRPM (%s):\n%w", srpmPath, err)
+				return
+			}
+
+			isToolchain := schedulerutils.IsReservedFile(rpmPath, toolchainRPMs)
+			if isToolchain {
+				rpmPath = convertToToolchainRpmPath(rpmPath, packagearch, toolchainDir)
+			}
+
 			providerPkgVer := &pkgjson.Package{
 				Provides:     newProviderVer[0],
 				SrpmPath:     srpmPath,
@@ -460,40 +568,23 @@ func parseProvides(rpmsDir, srpmPath string, list []string) (providerlist []*pkg
 	return
 }
 
-// parsePackageVersions takes a package name and splits it into a set of PackageVer structures.
+// parsePackageVersions takes a package string and splits it into a set of PackageVer structures.
 // Normally a list of length 1 is returned, however parsePackageVersions is also responsible for
 // identifying if the package name is an "or" condition and returning all options.
-func parsePackageVersions(packagename string) (newpkgs []*pkgjson.PackageVer, err error) {
-	const (
-		NameField      = iota
-		ConditionField = iota
-		VersionField   = iota
-	)
+func parsePackageVersions(packageString string) (newPackages []*pkgjson.PackageVer, err error) {
+	packageString = strings.TrimSpace(packageString)
 
-	packageSplit := strings.Split(packagename, " ")
-	err = minSliceLength(packageSplit, 1)
+	// If the first character of the packageString is a "(" then it's an "or" condition.
+	if packageString[0] == '(' {
+		return parseRichDependency(packageString)
+	}
+
+	pkgVer, err := pkgjson.PackageStringToPackageVer(packageString)
 	if err != nil {
 		return
 	}
 
-	// If first character of the packagename is a "(" then its an "or" condition
-	if packagename[0] == '(' {
-		return parseOrCondition(packagename)
-	}
-
-	newpkg := &pkgjson.PackageVer{Name: packageSplit[NameField]}
-	if len(packageSplit) == 1 {
-		// Nothing to do, no condition or version was found.
-	} else if packageSplit[ConditionField] != "or" {
-		newpkg.Condition = packageSplit[ConditionField]
-		newpkg.Version = packageSplit[VersionField]
-	} else {
-		// Replace the name with the first name that appears in (foo or bar)
-		substr := packageSplit[NameField][1:]
-		newpkg.Name = substr
-	}
-
-	newpkgs = append(newpkgs, newpkg)
+	newPackages = append(newPackages, pkgVer)
 	return
 }
 
@@ -511,10 +602,10 @@ func parsePackageVersionList(pkgList []string) (pkgVerList []*pkgjson.PackageVer
 	return
 }
 
-// condensePackageVersionArray deduplicates entries in an array of Package Versions
+// dedupPackageVersionArray deduplicates entries in an array of Package Versions
 // and represents double conditionals in a single PackageVersion structure.
-// If a non-blank package version is specified more than twice in a SPEC then return an error.
-func condensePackageVersionArray(packagelist []*pkgjson.PackageVer, specfile string) (processedPkgList []*pkgjson.PackageVer, err error) {
+// If a non-blank package version is specified more than twice, return an error.
+func dedupPackageVersionArray(packagelist []*pkgjson.PackageVer) (processedPkgList []*pkgjson.PackageVer, err error) {
 	for _, pkg := range packagelist {
 		nameMatch := false
 		for i, processedPkg := range processedPkgList {
@@ -535,7 +626,7 @@ func condensePackageVersionArray(packagelist []*pkgjson.PackageVer, specfile str
 					processedPkgList[i].SCondition = pkg.Condition
 					break
 				} else {
-					err = fmt.Errorf("spec (%s) attempted to set more than two conditions for package (%s)", specfile, processedPkg.Name)
+					err = fmt.Errorf("attempting to set more than two conditions for package (%s)", processedPkg.Name)
 					return
 				}
 			}
@@ -549,26 +640,63 @@ func condensePackageVersionArray(packagelist []*pkgjson.PackageVer, specfile str
 	return
 }
 
-// parseOrCondition splits a package name like (foo or bar) and returns both foo and bar as separate requirements.
-func parseOrCondition(packagename string) (versions []*pkgjson.PackageVer, err error) {
-	logger.Log.Warnf("'OR' clause found (%s), make sure both packages are available. Please refer to 'docs/how_it_works/3_package_building.md#or-clauses' for explanation of limitations.", packagename)
-	packagename = strings.ReplaceAll(packagename, "(", "")
-	packagename = strings.ReplaceAll(packagename, ")", "")
+// parseRichDependency splits a package name like '(foo or bar)' and returns both foo and bar as separate requirements.
+func parseRichDependency(richDependency string) (versions []*pkgjson.PackageVer, err error) {
+	const documentationHint = "Please refer to 'docs/how_it_works/3_package_building.md#rich-dependencies' for explanation of limitations"
 
-	packageSplit := strings.Split(packagename, " or ")
-	err = minSliceLength(packageSplit, 1)
+	// All single condition strings are surrounded by spaces to match full words.
+	for _, singleCondition := range unsupportedBooleanConditions {
+		if strings.Contains(richDependency, singleCondition) {
+			err = fmt.Errorf("found unsupported boolean condition '%s' inside '%s'. %s", singleCondition, richDependency, documentationHint)
+			return
+		}
+	}
+
+	conditionsCount := 0
+	// All single condition strings are surrounded by spaces to match full words.
+	for _, singleCondition := range supportedBooleanConditions {
+		conditionsCount += strings.Count(richDependency, singleCondition)
+	}
+	if conditionsCount > 1 {
+		err = fmt.Errorf("found more than one boolean condition inside '%s'. %s", richDependency, documentationHint)
+		return
+	}
+
+	richDependency = strings.ReplaceAll(richDependency, "(", "")
+	richDependency = strings.ReplaceAll(richDependency, ")", "")
+
+	packageStrings := []string{}
+	// All single condition strings are surrounded by spaces to match full words.
+	for _, singleCondition := range supportedBooleanConditions {
+		if strings.Contains(richDependency, singleCondition) {
+			packageStrings = strings.Split(richDependency, singleCondition)
+			break
+		}
+	}
+	err = minSliceLength(packageStrings, 2)
 	if err != nil {
 		return
 	}
 
-	versions = make([]*pkgjson.PackageVer, 0, len(packageSplit))
-	for _, condition := range packageSplit {
-		var parsedPkgVers []*pkgjson.PackageVer
-		parsedPkgVers, err = parsePackageVersions(condition)
+	switch {
+	case strings.Contains(richDependency, andCondition) || strings.Contains(richDependency, orCondition) || strings.Contains(richDependency, withCondition):
+		logger.Log.Warnf("Found a boolean condition '%s', make sure both packages are available. %s.", richDependency, documentationHint)
+	case strings.Contains(richDependency, ifCondition):
+		logger.Log.Warnf("Found a boolean condition '%s', make sure the packages on the left is available. %s.", richDependency, documentationHint)
+		packageStrings = []string{packageStrings[0]}
+	default:
+		err = fmt.Errorf("found a unsupported boolean condition inside '%s'. %s", richDependency, documentationHint)
+		return
+	}
+
+	versions = make([]*pkgjson.PackageVer, 0, len(packageStrings))
+	for _, packageString := range packageStrings {
+		pkgVer, err := pkgjson.PackageStringToPackageVer(packageString)
 		if err != nil {
-			return
+			return nil, err
 		}
-		versions = append(versions, parsedPkgVers...)
+
+		versions = append(versions, pkgVer)
 	}
 
 	return
@@ -605,4 +733,39 @@ func filterOutDynamicDependencies(pkgVers []*pkgjson.PackageVer) (filteredPkgVer
 	}
 
 	return
+}
+
+// convertToToolchainRpmPath updates the RPM path to point the the toolchain directory rather than the normal out/RPMs
+// directory
+func convertToToolchainRpmPath(currentRpmPath, arch, toolchainDir string) (toolchainPath string) {
+	rpmFileName := filepath.Base(currentRpmPath)
+	toolchainPath = filepath.Join(toolchainDir, arch)
+	toolchainPath = filepath.Join(toolchainPath, rpmFileName)
+	logger.Log.Debugf("Toolchain changing '%s' to '%s'.", currentRpmPath, toolchainPath)
+	return toolchainPath
+}
+
+func readBuildRequires(specFile, sourceDir, arch string, defines map[string]string) (result []*pkgjson.PackageVer, err error) {
+	const emptyQueryFormat = ``
+
+	queryResults, err := rpm.QuerySPEC(specFile, sourceDir, emptyQueryFormat, arch, defines, rpm.BuildRequiresArgument)
+	if err != nil {
+		return
+	}
+
+	result, err = parsePackageVersionList(queryResults)
+	if err != nil {
+		return
+	}
+
+	result, err = dedupPackageVersionArray(result)
+	if err != nil {
+		err = fmt.Errorf("failed to dedup build-time PackageVer array for spec (%s):\n%w", specFile, err)
+	}
+
+	return
+}
+
+func sendEmptyResult(results chan<- *parseResult, err error) {
+	results <- &parseResult{err: err}
 }
